% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{acquisitionMethods<-}
\alias{acquisitionMethods<-}
\title{acquisitionMethods<-}
\usage{
acquisitionMethods(object) <- value
}
\arguments{
\item{object}{Object to assign acquisition method names to an object.}

\item{value}{A character vector with acquisition method names.}
}
\value{
An object updated with method names.
}
\description{
acquisitionMethods<-
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
