% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{acquisitionMethods}
\alias{acquisitionMethods}
\title{acquisitionMethods}
\usage{
acquisitionMethods(object)
}
\arguments{
\item{object}{Object to collect method names used for data acquisition.}
}
\value{
A vector with method names.
}
\description{
acquisitionMethods
}
