% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_004_peakGrouping.R
\name{addAdjustedRetentionTime}
\alias{addAdjustedRetentionTime}
\title{addAdjustedRetentionTime}
\usage{
addAdjustedRetentionTime(object)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object.}
}
\description{
Function to add adjusted retention time information
to the scans slot of the \linkS4class{ntsData} object.
}
