% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_001_makeSetup.R
\name{addFiles}
\alias{addFiles}
\title{addFiles}
\usage{
addFiles(
  projPath = base::getwd(),
  newFiles = utils::choose.files(),
 
    copyFiles = utils::askYesNo("Do you want to copy the analysis files to the project directory?"),
  sampleInfo = base::data.frame(filePath = character(), sample = character(), group =
    character(), blank = character(), polarity = character(), date = character(), wdir =
    character()),
  date = base::Sys.Date(),
  groups = NULL,
  blanks = NULL,
  polarity = "positive"
)
}
\arguments{
\item{projPath}{The \code{path} where the project is located and files should be stored.}

\item{newFiles}{A list of analysis files to be added to the project. Defaults to a dialog to choose the new files.}

\item{copyFiles}{Logical. Parameter to decide whether the analysis files shall be copied to the project directory or remain at their original location. \code{TRUE} means files are copied and \code{FALSE} means files
 only remain in their original location. Defaults to a user prompt asking if files should be copied. As unique filenames are necessary the original files might have to be renamed. A second prompt is then given
 for confirmation.
might have to be renamed during the process. A confirmation prompt is}

\item{sampleInfo}{Already existing sampleInfo dataframe to which the new files are to be amended. A new dataframe will be created if no dataframe is given.}

\item{date}{The project date. The default is the system date generated by \code{\link[base]{Sys.time}}.}

\item{groups}{A vector with the identifier/name of each sample replicate group. If \code{NULL},
a grouping tentative will be made using the name of the MS files found in the given \code{projPath}.}

\item{blanks}{The name of the sample replicate group to be used for blank subtraction.
Different blank groups can be assigned to replicate sample groups by editing the \code{sampleInfo} in the output \code{setup} list.}

\item{polarity}{A vector specifying the polarity mode used in each MS file.
Possible values are \code{positive} or \code{negative} for positive and negative mode, respectively.}
}
\value{
Returns a dataframe containing the sample data for the analysis.
}
\description{
Add files to the sampleInfo object
}
