% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_001_makeSetup.R
\name{addMetadata}
\alias{addMetadata}
\title{addMetadata}
\usage{
addMetadata(sampleInfo, metadata)
}
\arguments{
\item{sampleInfo}{The sampleInfo \code{data.frame} object to be amended with metadata.}

\item{metadata}{A \code{data.frame} with the same number of rows as the sampleInfo contaninig 1 or more columns with metadata.
Note that the column names in the metadata \code{data.frame} will be used as classifiers.}
}
\value{
Returns an updated sampleInfo \code{data.frame} objects.
}
\description{
Adds additional information to the sampleInfo \code{data.frame}. For instance,
additional information could be concentration data or other known properties or classifiers for each sample in the sampleInfo \code{data.frame}.
The added metadata can then be used for inferential and statistical analyses.
}
