% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_000_suspectList.R
\name{addMS2Info}
\alias{addMS2Info}
\title{addMS2Info}
\usage{
addMS2Info(
  wfobj = NULL,
  sampleGroup = NULL,
  suspectList = NULL,
  param = MS2param(),
  updateRT = TRUE,
  updateIntControl = TRUE,
  save = TRUE
)
}
\arguments{
\item{wfobj}{An \linkS4class{ntsSuspectData} object}

\item{sampleGroup}{A character object with the name of a
representative sample replicate group to extract the MS2 information.}

\item{suspectList}{The \linkS4class{suspectList} object to add the MS2 information.}

\item{param}{The parameters to extract the MS2 from the \linkS4class{ntsSuspectData} object.}

\item{updateRT}{Logical, set to \code{TRUE} to update or add the retention time
to the \linkS4class{suspectList}.}

\item{updateIntControl}{Logical, set to \code{TRUE} to add the intensity
of each compound to the \emph{intControl} column in the \linkS4class{suspectList}.}

\item{save}{Logical, set to \code{TRUE} to update the csv file in the disk.
Optionally, a character object with the new name for the csv file.
Note, do not add the file extension to the name.}
}
\value{
A \linkS4class{suspectList} object with updated MS2 information
for each compound identified in the \linkS4class{ntsSuspectData} object.
}
\description{
addMS2Info
}
