% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_002_InspectRawData.R
\name{addRawInfo}
\alias{addRawInfo}
\title{addRawInfo}
\usage{
addRawInfo(object)
}
\arguments{
\item{ojbect}{An \linkS4class{ntsData} object.}
}
\value{
An \linkS4class{ntsData} object with raw information added
per sample in the samples slot.
}
\description{
Adds raw info for each sample (i.e., file)
in the samples slot of an \linkS4class{ntsData} object.
}
