% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_000_S4_parameterClasses.R
\docType{class}
\name{AlteredCameraParam-class}
\alias{AlteredCameraParam-class}
\title{AlteredCameraParam}
\value{
An \linkS4class{AlteredCameraParam} object containing parameters for
annotation of features in an \linkS4class{ntsData} object.
}
\description{
AlteredCameraParam
}
\section{Slots}{

\describe{
\item{\code{sigma}}{The multiplier of the standard deviation for grouping features
by retention time.}

\item{\code{perfwhm}}{Percentage of the overlapping FWHM to group features.}

\item{\code{cor_eic_th}}{Threshold for feature EIC correlation in each sample.}

\item{\code{calcCaS}}{Calculate correlation across samples.}

\item{\code{cor_exp_th}}{Threshold for intensity correlation across samples.}

\item{\code{calcIso}}{Logical, calculates based on isotopic information.}

\item{\code{pval}}{p-value threshold for testing correlation of significance.}

\item{\code{validateIsotopePatterns}}{Logical, set to \code{TRUE} for validating
the annotated isotopes with the \emph{kegg} database.}

\item{\code{ppmIsotopes}}{The expected mass deviation (in ppm) to find isotopes.}

\item{\code{mzabs}}{numeric.}

\item{\code{noise}}{numeric.}

\item{\code{searchAdducts}}{Logical, set to \code{TRUE} to screen for adducts after finding isotopes.}

\item{\code{ppmAdducts}}{The expected mass deviation (in ppm) to find adducts.}

\item{\code{extendedList}}{Logical, set to \code{TRUE} to use the extended list of
adducts. The default is \code{FALSE}.}
}}

