% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_005_annotation.R
\name{annotateFeatures}
\alias{annotateFeatures}
\title{annotateFeatures}
\usage{
annotateFeatures(
  obj = NULL,
  algorithm = NULL,
  param = NULL,
  excludeBlanks = FALSE,
  save = FALSE
)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object with features.}

\item{algorithm}{The algorithm for finding isotopes.
Currently, "alteredcamera" in the only possible value.}

\item{param}{The param used for annotation of isotopes and adducts.
See ?AlteredCameraParam for more information.}

\item{excludeBlanks}{Logical, set to \code{TRUE} for excluding
blank replicate groups from annotation.}

\item{save}{Logical, set to \code{TRUE} to save updated
\linkS4class{ntsData} object in the \strong{rdata} folder.
Note that \code{TRUE} overwrites the existing \linkS4class{ntsData} object.
Optionally, a character string can be given instead of \code{TRUE}
to be used as file name, avoiding overwriting.}
}
\value{
An \linkS4class{ntsData} object, containing components with
annotated features.
}
\description{
Group features into components according to co-elution and
EIC similarities. Then, it annotates isotopes and adducts in each component
per sample replicate group.
}
