% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{annotationParameters}
\alias{annotationParameters}
\title{annotationParameters}
\usage{
annotationParameters(object, algorithm, settings)
}
\arguments{
\item{object}{An object to add or get annotation parameter settings.}

\item{algorithm}{Algorithm for the parameters.}

\item{settings}{List of parameter settings.}
}
\value{
A list of parameters or the object with added parameters.
}
\description{
annotationParameters
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
