% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{blanks<-}
\alias{blanks<-}
\title{blanks<-}
\usage{
blanks(object) <- value
}
\arguments{
\item{object}{Object to assign blank sample replicate groups.}

\item{value}{A character vector with names of the sample replicate groups
to be used as blanks.}
}
\value{
An object updated with blank sample replicate groups.
}
\description{
blanks<-
}
