% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{blanks}
\alias{blanks}
\title{blanks}
\usage{
blanks(object)
}
\arguments{
\item{object}{Object to collect blank sample replicate groups.}
}
\value{
A vector with the blank sample replicate group names.
}
\description{
blanks
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
