% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_010_filters.R
\name{blankThreshold}
\alias{blankThreshold}
\title{blankThreshold}
\usage{
blankThreshold(obj, value = 3)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{value}{A numerical value to multiply the assigned blank intensity.}
}
\description{
blankThreshold
}
