% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_006_buildFeatureList.R
\name{buildFeatureList}
\alias{buildFeatureList}
\title{buildFeatureList}
\usage{
buildFeatureList(
  x = featData,
  xA = featComp,
  xPat = patData,
  snWindow = 240,
  save = TRUE,
  projPath = setup$projPath
)
}
\arguments{
\item{x}{A \linkS4class{XCMSnExp} object.}

\item{xA}{A list of \linkS4class{xsAnnotate} objects per replicate group.}

\item{xPat}{A \linkS4class{featureGroups} object converted by \code{\link{getPatData}}.}

\item{snWindow}{Time in seconds to expand the peak width for calculation of the signal-to-noise ratio.}

\item{save}{Logical, set to \code{TRUE} to save the generated \code{fl} object in the disk.}

\item{projPath}{The \code{projPath} directory as defined in the \code{setup} object.}
}
\value{
A \code{data.frame} with detailed information for each feature in the given objects.
}
\description{
Function to create a \code{data.frame} with detailed information for each feature in the given a \linkS4class{XCMSnExp} object object.
Optionally, a list of \linkS4class{xsAnnotate} objects per replicate group as obtained by \code{\link{makeFeatureComponents}} can be given.
The information from annotated isotopes and adducts for each component will be added to the \code{data.frame}.
}
\examples{



}
