% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_004_peakGrouping.R
\name{buildFeatureTable}
\alias{buildFeatureTable}
\title{buildFeatureTable}
\usage{
buildFeatureTable(object)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object.}
}
\value{
An \linkS4class{ntsData} object with updated features slot.
}
\description{
buildFeatureTable
}
