% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_003_peakPicking.R
\name{buildPeakList}
\alias{buildPeakList}
\title{buildPeakList}
\usage{
buildPeakList(obj)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object containing a \linkS4class{features}
object in the slot \code{patdata}.}
}
\value{
A data.frame containing information for peaks in each sample.
}
\description{
buildPeakList
}
