% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_003_peakPicking.R
\name{buildPeaksTable}
\alias{buildPeaksTable}
\title{buildPeaksTable}
\usage{
buildPeaksTable(object)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object
containing a \linkS4class{features}
object in the slot \code{pat}.}
}
\value{
A \link[data.table]{data.table} containing
information for peaks in each sample.
}
\description{
buildPeaksTable
}
