% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_008_FeaturesMetadata.R
\name{calculateFeaturesMetadata}
\alias{calculateFeaturesMetadata}
\title{calculateFeaturesMetadata}
\usage{
calculateFeaturesMetadata(obj, ID = NULL)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object containing a features.}

\item{ID}{Optional, vector with features IDs or indexes
for calculating metadata for the specified features.}
}
\value{
An \linkS4class{ntsData} object containing a features and peaks
quality metadata.
}
\description{
calculateFeaturesMetadata
}
