% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_009_featuresQuality.R
\name{calculateSNR}
\alias{calculateSNR}
\title{calculateSNR}
\usage{
calculateSNR(object, targets = NULL, rtExpand = 200)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object.}

\item{targets}{Optionally, a character vector with the \emph{id} of the features to calculate the sn ratio.}

\item{rtExpand}{The time, in seconds,
expansion before and after the time window of each feature
to extract centroids for estimation of the noise level.}
}
\value{
An \linkS4class{ntsData} object with the noise, noise sd and sn ratio columns ammended to the
features slot.
}
\description{
Calculates the signal-to-noise (sn) ratio for features in a given \linkS4class{ntsData} object
or specified via the argument \code{targets}. The noise is extimated by the maximum intensity of the background
centroids (i.e., traces belonging to the same mass bin as the peaks in the feature but taken from a wider and predefined
time window (\code{rtExpand}) before and after the feature limits. The sn is then calculated by
dividind the maximum intensity of the feature in the \linkS4class{ntsData} object with the estimated noise.
Centroids that belong to other peaks within the same mass and time windows are excluded before noise estimation.
}
