% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotMonitoring.R
\name{catPlot}
\alias{catPlot}
\title{catPlot}
\usage{
catPlot(
  Step,
  StepN,
  StepName,
  yUnit = "mz",
  size = NULL,
  constantLevel = 5000,
  ID = NULL
)
}
\arguments{
\item{Step}{A data frame with the categorized monitor data for a given step.}

\item{StepN}{The step number.}

\item{StepName}{The step name.}

\item{yUnit}{The unit to show in the y axis.
Possible values are "mz" and "rt".}

\item{size}{The size to plot the dots.}

\item{constantLevel}{The instensity variation threshold for constant features between steps.}

\item{ID}{Optional character vector with features ID to filter the process data.}
}
\value{
A ggplot with the features shown by categories according to the quantiles.
}
\description{
catPlot
}
