% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_012_monitoring.R
\name{catStep}
\alias{catStep}
\title{catStep}
\usage{
catStep(PL, IN, OUT, constantLevel = 5000)
}
\arguments{
\item{PL}{A features data frame.}

\item{IN}{The name of the sample replicate group as influent/input.}

\item{OUT}{The name of the sample replicate group as effluent/output.}
}
\value{
A data frame with features categorized as
removed (R), Lower (L), Constant (C), Higher (H) and New (N).
}
\description{
catStep
}
