% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_004_checkComponents.R
\name{checkComponents}
\alias{checkComponents}
\title{checkComponents}
\usage{
checkComponents(
  xA = featComp,
  replicateGroups = NULL,
  features = NULL,
  featData = featData,
  mz = NULL,
  ppm = 5,
  rt = NULL,
  rtWindow = 1,
  rtUnit = "min",
  comp = NULL,
  onlyAnnotated = FALSE,
  onlyRelated = TRUE,
  rGroupsAverage = FALSE,
  intval = "maxo"
)
}
\arguments{
\item{xA}{The list of \linkS4class{xsAnnotate} objects obtained via \code{\link{makeFeatureComponents}}.}

\item{replicateGroups}{The replicate groups name or number to filter the \code{xA} list,
which is named according to the experimental replicate groups as defined in \code{\link{makeSetup}}.}

\item{features}{Feature identifier/s.}

\item{featData}{The feature data to get the feature identifiers.
Can be a \linkS4class{XCMSnExp} or \linkS4class{featureGroups} object from \pkg{xcms} or \pkg{patRoon}, respectively.}

\item{mz}{The \emph{m/z} of interest. If \code{features} are specified \code{mz is not used}.}

\item{ppm}{The expected mass deviation to search for features of a given \code{mz}.}

\item{rt}{The expected retention time of the \emph{m/z} of interest, only used if \code{features} are not speficied.}

\item{rtWindow}{The expected retention time deviation for searching.}

\item{rtUnit}{The time unit used. Default is minutes.}

\item{comp}{The component numbers to plot. Only used if both \code{features} and \code{mz} are \code{NULL} (i.e. not specified).}

\item{onlyAnnotated}{Plots only annotated features.}

\item{onlyRelated}{Plots only relevant features (i.e. isotopes and addcuts) of the specified features.}

\item{rGroupsAverage}{Logical, set to \code{TRUE} for averaging the intensities of each sample acording the defined sample replicate groups.}

\item{intval}{The intensity value type. Default is "maxo", corresponding to the height.}
}
\value{
A data.frame with detailed information about a component of interest.
}
\description{
Inspect components by of given features, mz and retention time or component numbers of specified replicate groups
as defined in \code{\link{makeSetup}}.
}
\examples{



}
