% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_008_checkInternalStandards.R
\name{checkIS}
\alias{checkIS}
\title{checkIS}
\usage{
checkIS(
  x = NULL,
  xPat = NULL,
  polarity = "positive",
  screeningList = utils::choose.files(base::getwd(), "Select the QC screening list"),
  ppmWindow = 15,
  rtWindow = 30,
  plot = TRUE,
  save = TRUE,
  projPath = setup$projPath
)
}
\arguments{
\item{x}{A \linkS4class{XCMSnExp} object.}

\item{xPat}{Alternatively, a \linkS4class{featureGroups} object.}

\item{polarity}{The polarity for the search. Possible values are \code{positive} or \code{negative}.}

\item{screeningList}{A \code{data.frame} with details from each internal standard in the samples.
See details for more information about the required data.frame structure.}

\item{ppmWindow}{The mass deviation, in ppm, allowed to screen for internal standards.}

\item{rtWindow}{The retention time deviation, in seconds, allowed to screen for internal standards.}

\item{plot}{Logical, set to \code{TRUE} for plotting the results.}

\item{save}{Logical, set to \code{TRUE} for storing  the generated IS object in the rData folder.}

\item{projPath}{The \code{projPath} directory as defined in the \code{setup} object.}
}
\value{
A \code{list} containing the plots and a data.frame with the summary of the
retention time, mass and intensity deviations as well as the quality of the MS2 data for each internal standard found.
}
\description{
Evaluation of internal standards for samples within a given \linkS4class{XCMSnExp} or \linkS4class{featureGroups} object.
}
\details{
The \code{screeningList} template can be obtained as csv via \code{ntsIUTA::getScreeningListTemplate()}.
Add other details of the template.
}
\examples{



}
