% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_009_qualityControl.R
\name{checkQC}
\alias{checkQC}
\title{checkQC}
\usage{
checkQC(
  obj = NULL,
  targets = utils::choose.files(base::getwd(), "Select the QC screening list"),
  algorithmPeakPicking = NULL,
  algorithmMakeFeatures = NULL,
  paramPeakPicking = NULL,
  rtAlignment = TRUE,
  paramAlignment = NULL,
  paramGrouping = NULL,
  recursive = TRUE,
  paramFill = NULL,
  rtWindow = 30,
  ppm = 15,
  exportResults = FALSE,
  save = FALSE
)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object, containing QC samples.}

\item{targets}{A \code{data.frame} or a location for a csv
with details for each standard in the QC samples.
See details for more information about the required data.frame structure.}

\item{algorithmPeakPicking}{The algorithm to use for peak picking.
One of "xcms3" (the default), "xcms", "openms" or "envipick".}

\item{algorithmMakeFeatures}{The algorithm to use for peak alignment and grouping.
One of "xcms3" (the default), "xcms" or "openms".}

\item{paramPeakPicking}{List of arguments for the specified algorithm.
See \code{\link[patRoon]{findFeatures}} for more information.
For instance, for "xcms3" as \code{algorithm}, \code{param} should be given
with the method parameters for the peak finding.
See \href{https://rdrr.io/bioc/xcms/man/chromatographic-peak-detection.html}{\code{xcms::findChromPeaks}}
for more information.}

\item{rtAlignment}{Logical, set to \code{TRUE} (The default) for performing peak alignment before grouping.}

\item{paramAlignment}{Applicable for algorithm "xcms3" only,
the parameters for the chosen retention time alignment method.
See documentation of \code{\link[xcms]{adjustRtime}} for more information.
For alignment with the method \code{PeakGroups}, a list of length two
should be given, where the first element
is the pre-grouping parameters using \code{\link[xcms]{groupChromPeaks}}
and the second element the actual alignment parameters.}

\item{paramGrouping}{he parameters for the chosen grouping method.
See documentation of \code{\link[xcms]{groupChromPeaks}} or
\code{\link[patRoon]{groupFeatures}} for more information.}

\item{paramFill}{An object of class \code{FillChromPeaksParam}
or \code{ChromPeakAreaParam} containing the parameters
to apply the recursive integration.
#' See \code{?\link[xcms]{fillChromPeaks}} for more information.}

\item{rtWindow}{The retention time deviation, in seconds,
to screen for the QC target standards. The default is 30 seconds.}

\item{ppm}{The mass deviation, in ppm, to screen for the QC target standards.
The default is 15 ppm.}

\item{exportResults}{Logical, set to \code{TRUE} for exporting the results to the project folder.}

\item{save}{Logical, set to \code{TRUE} to save updated
\linkS4class{ntsData} object in the \strong{rdata} folder.
Note that \code{TRUE} overwrites the existing \linkS4class{ntsData} object.
Optionally, a character string can be given instead of \code{TRUE}
to be used as file name, avoiding overwriting.}

\item{recurvive}{Logical, set to \code{TRUE} for applying recursive
integration for samples with missing peaks.}
}
\value{
The \linkS4class{ntsData} with quality control data and
evaluation results added to the slot \code{QC}.
}
\description{
Function to check QC samples in an \linkS4class{ntsData} object.
}
\details{
The \code{screeningList} template can be obtained as csv via \code{ntsIUTA::getScreeningListTemplate()}.
Add other details of the template.
}
\examples{



}
