% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_002_InspectRawData.R
\name{clusterMSn}
\alias{clusterMSn}
\title{clusterMsn}
\usage{
clusterMSn(
  ids,
  msnList,
  clusteringMethod,
  clusteringUnit,
  clusteringWindow,
  mergeCEs,
  mergeBy,
  targets
)
}
\value{
A data table with clustered MSn data for given targets.
}
\description{
Function to cluster MSn data.
}
