% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{components}
\alias{components}
\title{components}
\usage{
components(object, ...)
}
\arguments{
\item{object}{An object to extract feature components.}

\item{...}{Other method dependent parameters.}
}
\value{
A data.frame with feature component details.
}
\description{
components
}
