% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_002_importRawData.R
\name{concatenate_OnDiskMSnExp}
\alias{concatenate_OnDiskMSnExp}
\title{concatenate_OnDiskMSnExp}
\usage{
concatenate_OnDiskMSnExp(...)
}
\arguments{
\item{...}{A list of OnDiskMSnExp objects to concatenate using \code{c()}.}
}
\value{
An OnDiskMSnExp object.
}
\description{
concatenate_OnDiskMSnExp
}
