% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{controlSamples<-}
\alias{controlSamples<-}
\title{controlSamples<-}
\usage{
controlSamples(object, ...) <- value
}
\arguments{
\item{object}{Object to collect the samples to be used as quality control.}

\item{value}{A character vector the names of the sample replicate group/s
to be used as quality control.}

\item{...}{Other argumments.}
}
\value{
An object with experimental samples assigned for quality control.
}
\description{
controlSamples<-
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
