% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{controlSamples}
\alias{controlSamples}
\title{controlSamples}
\usage{
controlSamples(object)
}
\arguments{
\item{object}{Object to collect the samples used for quality control.}
}
\value{
A vector with sample names used for quality control.
}
\description{
controlSamples
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
