% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{EICs}
\alias{EICs}
\title{EICs}
\usage{
EICs(object, ...)
}
\arguments{
\item{object}{Object to get extracted ion chromatograms (EICs).}
}
\value{
A \link[data.table]{data.table} with EICs in the object.
}
\description{
EICs
}
