% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{exportCSVs}
\alias{exportCSVs}
\title{exportCSVs}
\usage{
exportCSVs(object, path, ...)
}
\arguments{
\item{object}{An workflow object.}

\item{path}{The path to export the CSV files.}

\item{...}{Other parameters used in the production of CSV files.}
}
\value{
A set of CSVs saved into the given path.
}
\description{
exportCSVs
}
