% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{exportPlots}
\alias{exportPlots}
\title{exportPlots}
\usage{
exportPlots(object, path, ...)
}
\arguments{
\item{object}{A workflow object.}

\item{path}{The path to export the plots.}

\item{...}{Other parameters used in the production of plots.}
}
\value{
A set of plots saved into the given path.
}
\description{
exportPlots
}
