% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_002_InspectRawData.R
\name{extractEICs}
\alias{extractEICs}
\title{extractEICs}
\usage{
extractEICs(
  object = NULL,
  samples = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60
)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object.}

\item{samples}{A numeric or character vector with the index or names
of the sample/s to extract the data, respectively.
When \code{NULL}, all the samples in the \code{object} are used.}

\item{mz}{A numeric vector with target \emph{m/z} values to extract EICs.
Alternatively, a two columns data.table or data.frame with
minimum and maximum (in this order) \emph{m/z} values can be used instead of exact \emph{m/z}.
Note that for the latter, the argumment \code{ppm} is ignored.}

\item{ppm}{A numeric vector of length one with the mass deviation, in ppm, to extract the data for the EIC.}

\item{rt}{A numeric vector with retention time values, in seconds, to extract EICs.
When \code{NULL}, the complete retention time window in the sample (i.e., file) is used.
Alternatively, a two columns data.table or data.frame with
minimum and maximum (in this order) retention time values can be used instead of exact retention times.
Note that for the latter, the argumment \code{sec} is ignored.
Note that the legnth of the \code{rt} vector of number of rows of the \code{rt} table
should be equal to the length or number of rows of \code{mz}.}

\item{sec}{A numeric vector of length one with the time deviation, in seconds, to extract the data for the EIC.}
}
\value{
A \code{data.table} with the columns
\code{sample}, \code{replicate}, \code{id}, \code{rt}, and \code{intensity}
representing the sample index (i.e., file), the sample replicate name,
the EIC target id, the retention time and the sum of the intensities
for the collected \emph{m/z} in each spectrum (i.e., MS scan), respectively.
}
\description{
Extracts MS1 level ion chromatograms (EICs) from raw data of a specified \emph{m/z}
and retention time pair/s.
}
