% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_001_supportTools.R
\name{extractMS2}
\alias{extractMS2}
\title{extractMS2}
\usage{
extractMS2(obj = NULL, param = paramMS2())
}
\arguments{
\item{obj}{A \linkS4class{featureGroups} or an \linkS4class{ntsData} object to
extract MS2 data according to percursor ions.}

\item{param}{A \linkS4class{paramMS2} object with parameters for MS2 extraction.}
}
\value{
A \linkS4class{MSPeakLists} when \code{obj} is a \linkS4class{featureGroups}
or an \linkS4class{ntsData} updated with MS2 data
when \code{obj} is an \linkS4class{ntsData} object.
}
\description{
extractMS2
}
