% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_002_InspectRawData.R
\name{extractMSn}
\alias{extractMSn}
\title{extractMSn}
\usage{
extractMSn(
  object = NULL,
  samples = NULL,
  level = 2,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  sec = 60,
  algorithm = NA_character_,
  settings = NULL
)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object.}

\item{samples}{A numeric or character vector with the index or names
of the sample/s to extract the data, respectively.
When \code{NULL}, all the samples in the \code{object} are used.}

\item{level}{A numeric vector with length 1 to defined the MS level.
Currently, only level 2, corresponding to MS/MS, is possible.}

\item{mz}{A numeric vector with isolation target \emph{m/z} values to extract MSn.
Alternatively, a two columns data.table or data.frame with
minimum and maximum (in this order) \emph{m/z} values can be used instead of exact \emph{m/z}.
Note that for the latter, the argumment \code{ppm} is ignored.}

\item{ppm}{A numeric vector of length one with the mass deviation, in ppm, to screen for isolation targets.}

\item{rt}{A numeric vector with retention time values, in seconds, to extract MSn.
When \code{NULL}, the complete retention time window in the sample (i.e., file) is used.
Alternatively, a two columns data.table or data.frame with
minimum and maximum (in this order) retention time values can be used instead of exact retention times.
Note that for the latter, the argumment \code{sec} is ignored.
Note that the legnth of the \code{rt} vector of number of rows of the \code{rt} table
should be equal to the length or number of rows of \code{mz}.}

\item{sec}{A numeric vector of length one with the time deviation, in seconds, to screen for isolation targets.}

\item{algorithm}{A character vector of length one with the algorithm used to extract and cluster MS2 data.}

\item{settings}{A list of parameters settings.}
}
\value{
A \code{data.table} with the columns
\code{sample}, \code{replicate}, \code{id}, \code{voltage}, \code{mz}, \code{intensity} and \code{precursor}
representing the sample name (i.e., file), the sample replicate name,
the isolation target id, the collision energy applied, the \emph{m/z}, the intensity and the presence of the precursor
(i.e., \emph{m/z} matching the isolation target), respectively.
}
\description{
Extracts MSn spectra from defined isolated targets
defined by \emph{m/z} and retention time, including the respective deviations.
}
