% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{fillingParameters}
\alias{fillingParameters}
\title{fillingParameters}
\usage{
fillingParameters(object, algorithm, settings)
}
\arguments{
\item{object}{An object to add or get filling parameter settings.}

\item{algorithm}{Algorithm for the parameters.}

\item{settings}{List of parameter settings.}
}
\value{
A list of parameters or the object with added parameters.
}
\description{
fillingParameters
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
