% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_005_peakFilling.R
\name{fillingSettingsDefaultXCMS}
\alias{fillingSettingsDefaultXCMS}
\title{fillingSettingsDefaultXCMS}
\usage{
fillingSettingsDefaultXCMS()
}
\value{
An \linkS4class{ntsSettings} object containing parameters for
extraction of MS2 spectra of given precursor ions in an \linkS4class{ntsData} object.
}
\description{
fillingSettingsDefaultXCMS
}
