% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{fillMissingParameters}
\alias{fillMissingParameters}
\title{fillMissingParameters}
\usage{
fillMissingParameters(object, algorithm, param)
}
\arguments{
\item{object}{An object to add or get fill missing parameters.}

\item{...}{Other method dependent parameters.}
}
\value{
A list of parameters or the object with added parameters.
}
\description{
fillMissingParameters
}
