% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_007_filter.R
\name{filterBlank}
\alias{filterBlank}
\title{filterBlank}
\usage{
filterBlank(obj, blankMultiplier = 3)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{blankMultiplier}{A numerical value to multiply the assigned blank intensity.}
}
\value{

}
\description{
filterBlank
}
