% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_001_supportTools.R
\name{filterFeatureGroups}
\alias{filterFeatureGroups}
\title{filterFeatureGroups}
\usage{
filterFeatureGroups(x, i)
}
\arguments{
\item{x}{A \linkS4class{featureGroups} object.}

\item{i}{The indices or the sample names to subset \code{x}.}
}
\value{
A subset of the \linkS4class{featureGroups} object.
}
\description{
filterFeatureGroups
}
