% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_007_filterFeatures.R
\name{filterFeatures}
\alias{filterFeatures}
\title{filterFeatures}
\usage{
filterFeatures(obj, ...)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{...}{A sequence of filter arguments to applied to the features data frame.
The filter sequence order is user defined and is applied as given in the arguments.
See details for possible filter arguments.}
}
\value{
An \linkS4class{ntsData} object with filtered features annotated in the features slot.
}
\description{
filterFeatures
}
\details{
The available filters are as folllows:
\itemize{
 \item \code{filterMinInt} Filter features below a minimum intensity threshold. For example, filterMinInt = 3000, removing features below 3000 counts.
 \item \code{filterBlank} Filter features that are not more intense than a defined multiplier times the assigned blank intensity. For example, filterBlank = 3, removing features that are not higher than 3 times the blank intensity.
 \item \code{filterSNR} Filter features below a minimum signal-to-noise ratio (SNR) threshold. For example, filterSNR = 3, filtering features below a SNR of 3.
 \item \code{filterMinReplicateAbundance} Filter features that are not present with at least a specified frequency in one sample replicate group. For example, filterMinReplicateAbundance = 2, meaning that featues should be at least represented in two samples of at least one sample replicate group.
 \item \code{filterIntSD} Filter features based on a maximum standard deviation (SD) among replicate samples. The SD should be at least below the defined threshold, as percentage, in at least one sample replicate group. For example, filterIntSD = 30, filtering features that do not have the SD below 30% in at least one sample replicate group.
}
}
