% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_001_supportTools.R
\name{filterFileFaster}
\alias{filterFileFaster}
\title{filterFileFaster}
\usage{
filterFileFaster(x, i)
}
\arguments{
\item{x}{An \linkS4class{OnDiskMSnExp} object.}

\item{i}{The indices or names of the samples to keep.}
}
\value{
The sub-setted \linkS4class{OnDiskMSnExp} object.
}
\description{
Filter files (i.e., samples) but it does not redo feature list.
Useful for plotting data or extracting EICs from certain samples.
}
