% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_007_filter.R
\name{filterIntSD}
\alias{filterIntSD}
\title{filterIntSD}
\usage{
filterIntSD(obj, sdThresholdPerc = 22)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{sdThresholdPerc}{A numerical value set at the desired sd percentage maximum.}
}
\value{

}
\description{
filterIntSD
}
