% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_007_filter.R
\name{filterMinInt}
\alias{filterMinInt}
\title{filterMinInt}
\usage{
filterMinInt(obj, intensityThreshold = 500)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{intensityThreshold}{A numerical value set at the desired minimum intensity for features.}
}
\value{

}
\description{
filterMinInt
}
