% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_007_filter.R
\name{filterMinReplicateAbundance}
\alias{filterMinReplicateAbundance}
\title{filterMinReplicateAbundance}
\usage{
filterMinReplicateAbundance(obj, ReplicateThreshold = 2)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{ReplicateThreshold}{A numerical value set at the desired minimum replicates.}
}
\value{

}
\description{
filterMinReplicateAbundance
}
