% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_003_peakPicking.R
\name{filterPeaks}
\alias{filterPeaks}
\title{filterPeaks}
\usage{
filterPeaks(
  peaks = peaks,
  fileIndex = NULL,
  mz = NULL,
  ppm = NULL,
  rt = NULL,
  rtWindow = NULL,
  rtUnit = "sec",
  absMinIntensity = NULL,
  chromWidthRange = NULL,
  negate = FALSE,
  dataframe = TRUE
)
}
\arguments{
\item{mz}{The target emph{mz} or emph{mz} range, which should be given as a numeric vector with length 2.}

\item{ppm}{The mass deviation in ppm from the target emph{mz} to get peaks.}

\item{rt}{The target retention time to get peaks.}

\item{rtWindow}{The retention time deviation or the range (given as a numeric vector with length 2) to get peaks.}

\item{rtUnit}{The unit of the time arguments. Possible values are "sec" (the default) and "min".}

\item{absMinIntensity}{The absolute minimum intensity to obtain peaks.}

\item{chromWidthRange}{The min and max limits for chromatographic peak width given as a numeric vector with length 2.
Max can be \code{Inf} to specify no maximum limit.}

\item{negate}{Logical, set to \code{TRUE} to apply the opposite of the defined filters. The default is \code{FALSE}.}

\item{dataframe}{Logical, set to \code{TRUE} to return a data.frame instead of a \code{features} object. Usefull for inspection.}

\item{obj}{A \linkS4class{ntsData} object containing peaks.}
}
\value{
A filtered \linkS4class{features} object or a \code{data.frame} when dataframe argument is set to \code{TRUE}.
}
\description{
Filter peaks from a \linkS4class{ntsData} object based on
file, retention time and emph{m/z}.
See \code{\link[patRoon]{filter}} for more information about the arguments.
}
\examples{

}
\references{
\insertRef{Helmus2021}{ntsIUTA}
}
