% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_007_filter.R
\name{filterSNR}
\alias{filterSNR}
\title{filterSNR}
\usage{
filterSNR(obj, snRatio = 5)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{snRatio}{A numerical value set at the desired signal-to-noise ratio for features.}
}
\value{

}
\description{
filterSNR
}
