% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_007_peakAnnotationAlteredFromCAMERA.R
\name{findIsotopesForPSAltered}
\alias{findIsotopesForPSAltered}
\title{findIsotopesForPSAltered}
\usage{
findIsotopesForPSAltered(
  peakIndeces = peakIndeces,
  mzValues = mzValues,
  intValues = intValues,
  sdValues = sdValues,
  snValues = snValues,
  maxcharge = maxcharge,
  devppm = devppm,
  validateIsotopePatterns = TRUE
)
}
\arguments{
\item{peakIndeces}{Indices of peaks.}

\item{mzValues}{\emph{m/z} mzvalues of peaks.}

\item{intValues}{Intensity values of peaks.}

\item{sdValues}{Intensity standard deviation of the peaks.}

\item{snValues}{Signal-to-noise ration of the peaks.}

\item{maxcharge}{Maximum charge allowed.}

\item{devppm}{Maximum mass deviation, in ppm.}

\item{validateIsotopePatterns}{Logical, when \code{TRUE} applies validation using the kegg library.}
}
\value{
A list of isotopic annotation for the given peaks.
}
\description{
Function used within \code{FindIsotopesWithValidationAltered}.
}
\references{
\insertRef{CAMERA}{ntsIUTA}
}
