% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_005_isotopesAlteredFromCAMERA.R
\name{FindIsotopesWithValidationAltered}
\alias{FindIsotopesWithValidationAltered}
\title{FindIsotopesWithValidationAltered}
\usage{
FindIsotopesWithValidationAltered(
  object,
  featData = featData,
  sampleidxs = sampleidxs,
  maxcharge = 3,
  ppm = 50,
  mzabs = 0.01,
  noise = NULL,
  intval = "maxo",
  validateIsotopePatterns = TRUE
)
}
\arguments{
\item{object}{An \linkS4class{xsAnnotate} object.}

\item{featData}{An \linkS4class{XCMSnExp} object containing features.}

\item{sampleidxs}{The index of samples for the replicate groups of interest.}

\item{maxcharge}{Maximum number of possible charges.}

\item{ppm}{The mass deviation, in ppm, allowed for searching isotopes within each component.}

\item{mzabs}{The mass deviation, in \emph{m/z}, allowed for searching isotopes within each component.}

\item{noise}{The overall noise level. When given the sn is calculated based on the given noise level.}

\item{intval}{The intensity value to check isotopes. Default and recommended is "maxo", which is the peak height.}

\item{validateIsotopePatterns}{Logical, set to \code{TRUE} to validate isotopes using the kegg library.}
}
\value{
An \linkS4class{xsAnnotate} object with annotated isotopes.
}
\description{
Altered version of the \code{findIsotopesWithValidation} function from \code{CAMERA} package.
}
\examples{



}
\references{
\insertRef{CAMERA}{ntsIUTA}
}
