% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_006_peakFragments.R
\name{fragmentSettingsDefault}
\alias{fragmentSettingsDefault}
\title{fragmentSettingsDefault}
\usage{
fragmentSettingsDefault()
}
\value{
An \linkS4class{ntsSettings} object containing parameters for
extraction of MS2 spectra of given precursor ions in an \linkS4class{ntsData} object.
}
\description{
fragmentSettingsDefault
}
