% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_006_peakFragments.R
\name{generateMS2}
\alias{generateMS2}
\title{generateMS2}
\usage{
generateMS2(object, algorithm = NA_character_, settings = NULL)
}
\arguments{
\item{object}{An \linkS4class{ntsData} or a \linkS4class{featureGroups} object to
extract MS2 data according to percursor ions.}

\item{algorithm}{A character vector with the algorithm to generate MS2 data.}

\item{settings}{A list of settings to use for extraction of MS2 data.}
}
\value{
A \linkS4class{MSPeakLists} when \code{object} is a \linkS4class{featureGroups}
or an \linkS4class{ntsData} with MS2 data added to the features data.table
when \code{object} is an \linkS4class{ntsData} object.
}
\description{
generateMS2
}
