% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_001_supportTools.R
\name{getColors}
\alias{getColors}
\title{getColors}
\usage{
getColors(x, which = c("samples", "samplegroups", "groups"))
}
\arguments{
\item{x}{An \linkS4class{OnDiskMSnExp} object with one or more files or the number of colors to be produced.}

\item{which}{Possible entries are \code{groups}, \code{samples} and \code{perSample} for getting
group, sample group or individual sample colors, respectively.}
}
\value{
A vector of colors names according to the given \linkS4class{OnDiskMSnExp} object.
}
\description{
getColors
}
