% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_CorReplicates.R
\name{getCorReplicates}
\alias{getCorReplicates}
\title{getCorReplicates}
\usage{
getCorReplicates(obj, binsize = 2, exportResults = FALSE)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object with more than one file.}

\item{binsize}{The bin size to apply to the raw data. The default is 2.}

\item{exportResults}{Logical, set to \code{TRUE} for exporting the results to the project folder.}
}
\value{
A data frame with the minimum, average and respective sd of the correlation between
sample replicate groups.
}
\description{
getCorReplicates
}
