% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_002_InspectRawData.R
\name{getRawInfo}
\alias{getRawInfo}
\title{getRawInfo}
\usage{
getRawInfo(object, samples = NULL)
}
\arguments{
\item{samples}{A numeric or character vector with the index or name of
samples to collect raw inforamtion, respectively. The default is \code{NULL}
which considers all samples in the \code{object}.}

\item{ojbect}{An \linkS4class{ntsData} object.}
}
\value{
A \link[data.table]{data.table} with raw information for each specified file.
}
\description{
Get raw information for each sample (i.e., file) in an
\linkS4class{ntsData} object.
}
