% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_001_supportTools.R
\name{getScreeningListTemplate}
\alias{getScreeningListTemplate}
\title{getScreeningListTemplate}
\usage{
getScreeningListTemplate(projPath = setup$projPath)
}
\arguments{
\item{projPath}{The project folder location. Default is \code{setup$projPath}.}
}
\value{
Pastes a template .csv file of the screeningList into the projPath.
}
\description{
getScreeningListTemplate
}
