% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_000_suspectList.R
\name{getSuspectList}
\alias{getSuspectList}
\title{getSuspectList}
\usage{
getSuspectList(path = NULL, comment = NA_character_)
}
\arguments{
\item{path}{The complete path to load the csv file of the list of suspects.
Note that the \emph{suspectListTemplate} from \pkg{ntsIUTA}
should be used to build the suspect list without changing the columns.
The template can be obtained with \link{getSuspectListTemplate}.}

\item{comment}{Optional comment for the suspect list.}
}
\value{
An \linkS4class{suspectList} object containing information of
suspect compounds used for screening workflows within \pkg{ntsIUTA}.
}
\description{
getSuspectList
}
