% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_000_S4_suspectList.R
\name{getSuspectListTemplate}
\alias{getSuspectListTemplate}
\title{getSuspectListTemplate}
\usage{
getSuspectListTemplate(saveTo = NULL)
}
\arguments{
\item{saveTo}{The path to save the \emph{suspectListTemplate.csv} file.}
}
\value{
A csv template for building a \linkS4class{suspectList} object.
}
\description{
getSuspectListTemplate
}
