% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{hasAdjustedRetentionTime}
\alias{hasAdjustedRetentionTime}
\title{hasAdjustedRetentionTime}
\usage{
hasAdjustedRetentionTime(object)
}
\arguments{
\item{object}{A object to check for existence of adjusted retention time.}
}
\value{
A logical value.
}
\description{
hasAdjustedRetentionTime
}
