% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_002_importRawData.R
\name{importRawData}
\alias{importRawData}
\title{importRawData}
\usage{
importRawData(
  obj = NULL,
  rtFilter = NULL,
  timeUnit = "min",
  msLevel = c(1, 2),
  centroidedData = TRUE,
  removeEmptySpectra = TRUE,
  save = FALSE
)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{rtFilter}{A numeric vector with length 2 defining the minimum
and maximum chromatographic retention time for the listed MS files.}

\item{timeUnit}{The unit of the \code{rtFilter}.
Possible values are \code{min} (the default) and \code{sec}.}

\item{msLevel}{The MS dimensions for the rtFilter to be applied.
The default is both MS1 and MS2 using \code{c(1,2)}.}

\item{centroidedData}{Logical, set to \code{TRUE} for MS files
with centroided data or \code{FALSE} for profile data.
\code{NA} will collect all the data from the MS files.}

\item{removeEmptySpectra}{Logical, set to TRUE if empty spectra should be removed.
It is recommended to remove empty spectra as it may cause issues during creation of features.}

\item{save}{Logical, set to \code{TRUE} to save updated
\linkS4class{ntsData} object in the \strong{rdata} folder.
Note that \code{TRUE} overwrites the existing \linkS4class{ntsData} object.
Optionally, a character string can be given instead of \code{TRUE}
to be used as file name, avoiding overwriting.}
}
\value{
The \linkS4class{ntsData} object including a standard
\linkS4class{OnDiskMSnExp} object in the MSnExp slot. Note, that
the \linkS4class{OnDiskMSnExp} object can also be used within
the workflow of \pkg{Bioconductor} packages.
}
\description{
Function to import MS data from the MS files listed in
the \linkS4class{ntsData} object. Files should contain centroided spectra.
The function \code{\link[MSnbase]{readMSData}} from the
\code{MSnbase} package is used to read the MS files.
}
\examples{
path <- system.file(package = "ntsIUTA", dir = "extdata")
dt <- setupProject(path = path, save = FALSE)
dt <- importRawData(obj[1], save = FALSE, centroidedData = TRUE)

}
\references{
\insertRef{MSnbase1}{ntsIUTA}
\insertRef{MSnbase2}{ntsIUTA}
}
