% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{isData-class}
\alias{isData-class}
\title{isData}
\value{
A \linkS4class{qcData} object to store the QC parameters, data and results
from an \linkS4class{ntsData} object.
}
\description{
isData
}
\section{Slots}{

\describe{
\item{\code{targets}}{A \linkS4class{suspectList} object with the QC target compounds.}

\item{\code{ppm}}{The mass window, in ppm, to look for target compounds.}

\item{\code{rtWindow}}{The retention time window, in seconds, to look for target compounds.}

\item{\code{recoveryFrom}}{A character object with the name of the sample replicate group
used for recovery evaluation.}

\item{\code{results}}{A data frame with the QC results summary.}
}}

