% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_002_InspectRawData.R
\name{loadEICs}
\alias{loadEICs}
\title{loadEICs}
\usage{
loadEICs(spectra, rtMins, rtMaxs, mzMins, mzMaxs)
}
\arguments{
\item{spectra}{The spectra data to build the EICs.}

\item{rtMins}{A numeric vector with the minimum retention times of EICs to build.}

\item{rtMaxs}{A numeric vector with the maximum retention times of EICs to build.}

\item{mzMins}{A numeric vector with the minimum mass-to-charge ratio of EICs to build.}

\item{mzMaxs}{A numeric vector with the maximum mass-to-charge ratio of EICs to build.}
}
\description{
Loads extracted ion chromatograms
from pre-loaded spectra data.
The function is imported from the \pkg{patRoon}.
}
