% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_006_peakFragments.R
\name{loadMS2}
\alias{loadMS2}
\title{loadMS2}
\usage{
loadMS2(object, algorithm = NA_character_, settings = NULL, save = FALSE)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object.}

\item{algorithm}{A character vector with the algorithm to generate MS2 data.}

\item{settings}{A list of settings to use for extraction of MS2 data.}

\item{save}{Logical, set to \code{TRUE} to save updated
\linkS4class{ntsData} object in the \strong{rdata} folder.
Note that \code{TRUE} overwrites the existing \linkS4class{ntsData} object.
Optionally, a character string can be given instead of \code{TRUE}
to be used as file name, avoiding overwriting.}
}
\value{
An \linkS4class{ntsData} with MS2 data added to the \code{ms2} slot.
}
\description{
Function to load MS2 data from features in the feature table
of a \linkS4class{ntsData} object. The MS2 data is added to the \code{ms2}
slot of the \linkS4class{ntsData} object. If the \code{algorithm} and
\code{settings} arguments are not given, the fucntion uses the parameters in the
\linkS4class{ntsData} object.
}
