% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{loadParameters}
\alias{loadParameters}
\title{loadParameters}
\usage{
loadParameters(object, ...)
}
\arguments{
\item{object}{An object to add parameter settings.}

\item{...}{Other arguments.}
}
\value{
The object with added parameters.
}
\description{
loadParameters
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
