% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_002_InspectRawData.R
\name{loadSpectra}
\alias{loadSpectra}
\title{loadSpectra}
\usage{
loadSpectra(path, rtRange = NULL, verbose = TRUE, cacheDB = NULL)
}
\arguments{
\item{rtRange}{A vector with length two specifying the time window
(i.e., mininum and maximum) to extract the spectra.}

\item{verbose}{Logical, set to \code{TRUE} to print processing information.}

\item{cacheDB}{The database for caching information.
Use openCacheDBScope() to set the cacheDB and enable integration within \pkg{patRoon}.}

\item{filepath}{A charcater string with the complete file path to load the spectra.}
}
\description{
Loads spectra data for a given file.
The function is imported from the \pkg{patRoon}.
}
