% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_005_makeFeatureComponents.R
\name{makeFeatureComponents}
\alias{makeFeatureComponents}
\title{makeFeatureComponents}
\usage{
makeFeatureComponents(
  featData = featData,
  polarity = "positive",
  sigma = 5,
  perfwhm = 0.5,
  cor_eic_th = 0.85,
  cor_exp_th = 0.75,
  pval = 0.05,
  calcCaS = TRUE,
  calcIso = TRUE,
  ppmIsotopes = 50,
  mzabs = 0.01,
  noise = 350,
  validateIsotopePatterns = TRUE,
  searchAdducts = TRUE,
  ppmAdducts = 5,
  extendedList = FALSE,
  excludeBlanks = TRUE,
  blankGroups = "Blank",
  save = TRUE,
  projPath = setup$projPath
)
}
\arguments{
\item{featData}{An \linkS4class{XCMSnExp} object containing features.}

\item{polarity}{The polarity of the replicate groups. Possible values are \code{positive} or \code{negative}.}

\item{sigma}{The multiplier of the standard deviation for grouping features by retention time.}

\item{perfwhm}{Percentage of the width of the FWHM.}

\item{cor_eic_th}{Minimum correlation index (from 0 to 1) for the EICs of features within the same sample.}

\item{cor_exp_th}{Minimum correlation index (from 0 to 1) for the EICs of features across samples within the replicate group.}

\item{pval}{p-value threshold for testing correlation of significance.}

\item{calcCaS}{Logical, set to \code{TRUE} to calculate correlation accross samples. Deafault is \code{TRUE}.}

\item{calcIso}{Logical, set to \code{TRUE} to include isotope detection informationen for graph clustering. Deafault is \code{FALSE}.}

\item{ppmIsotopes}{The expected mass deviation to find isotopes.}

\item{mzabs}{The expected deviation of the \emph{m/z} to find isotopes.}

\item{noise}{The extimated intensity threshold for the noise level used for find isotopes.}

\item{validateIsotopePatterns}{Logical, set to \code{TRUE} for validating the annoatated isotopes with the kegg database.}

\item{searchAdducts}{Logical, set to \code{TRUE} to screen for adducts after finding isotopes.}

\item{ppmAdducts}{The expected mass deviation to find adducts.}

\item{extendedList}{Logical, set to \code{TRUE} to use the extended list of adducts. The default (\code{FALSE} uses a shorter list.)}

\item{excludeBlanks}{Set to \code{TRUE} to not screen blank samples for isotopes and adducts.
This option is intended for saving processing time as the replicate blank groupes are expected to be removed before final feature list creation.}

\item{blankGroups}{A character vector with the name of the blank replicate groups in the given \code{featData} object.
For simplification use \code{base::unique(sampleInfo$blank)}.}

\item{save}{Logical, set to \code{TRUE} to save the generated \code{list} of \linkS4class{xsAnnotate} objects in the disk.}

\item{projPath}{The \code{projPath} directory as defined in the \code{setup} object.}
}
\value{
A \code{list} with an \linkS4class{xsAnnotate} object per replicate group as defined in the given \linkS4class{XCMSnExp} object.
}
\description{
Uses the \pkg{CAMERA} package to find isotopes by grouping features over the retention time
and extracted ion chromotogram (EIC) for the given deviations. The isotopes are found for each replicate group.
}
\examples{



}
\references{
\insertRef{CAMERA}{ntsIUTA}
\insertRef{xcms1}{ntsIUTA}
\insertRef{xcms2}{ntsIUTA}
\insertRef{xcms3}{ntsIUTA}
}
