% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_004_makeFeatures.R
\name{makeFeatures}
\alias{makeFeatures}
\title{makeFeatures}
\usage{
makeFeatures(
  obj = NULL,
  algorithm = NULL,
  paramGrouping = NULL,
  paramFill = NULL,
  save = TRUE
)
}
\arguments{
\item{obj}{A \linkS4class{ntsData} object containing peaks.}

\item{algorithm}{The algorithm to use for peak alignment and grouping.}

\item{paramGrouping}{The parameters for the chosen grouping method.
See \code{\link[patRoon]{groupFeatures}} for more information.}

\item{paramFill}{A list of parameters for recursive integration of missing peaks.
See \code{\link[xcms]{fillChromPeaks}} for more information.}

\item{save}{Logical, set to \code{TRUE} to save updated
\linkS4class{ntsData} object in the \strong{rdata} folder.
Note that \code{TRUE} overwrites the existing \linkS4class{ntsData} object.
Optionally, a character string can be given instead of \code{TRUE}
to be used as file name, avoiding overwriting.}
}
\value{
An \linkS4class{ntsData} object
containing peaks aligned and grouped as features across samples accessible
in the slot \code{features}.

 @references
\insertRef{Helmus2021}{ntsIUTA}
\insertRef{xcms1}{ntsIUTA}
\insertRef{xcms2}{ntsIUTA}
\insertRef{xcms3}{ntsIUTA}
}
\description{
Grouping and alignment of peaks across samples.
The peak grouping uses the \pkg{patRoon} package and
the following algorithms are possible: "xcms3", "xcms", "openms".
See ?\pkg{patRoon} for further information.
When fill missing parameters (paramFill) are defined
a recursive integration for samples with missing peaks
can be applied through the \pkg{xcms} package.
}
