% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_001_makeSetup.R
\name{makeSetup}
\alias{makeSetup}
\title{makeSetup}
\usage{
makeSetup(
  projPath = utils::choose.dir(base::getwd(), "Select or create a project folder"),
  date = base::Sys.Date(),
  groups = NULL,
  blanks = NULL,
  polarity = "positive",
  convertFiles = FALSE,
  convertFrom = NULL,
  convertToCentroid = TRUE,
  save = TRUE,
  makeNewProject = FALSE
)
}
\arguments{
\item{projPath}{The directory for the project. The default is the \code{utils::choose.dir()} to select or create a folder.
Note that the function will look into subfolders for mzML files or specified raw files when \code{convertFiles} is set to \code{TRUE}.}

\item{date}{The project date. The default is the system date generated by \code{\link[base]{Sys.time}}.}

\item{groups}{A vector with the identifier/name of each sample replicate group. If \code{NULL},
a grouping tentative will be made using the name of the MS files found in the given \code{projPath}.}

\item{blanks}{The name of the sample replicate group to be used for blank subtraction.
Different blank groups can be assigned to replicate sample groups by editing the \code{sampleInfo} in the output \code{setup} list.}

\item{polarity}{A vector specifying the polarity mode used in each MS file.
Possible values are \code{positive} or \code{negative} for positive and negative mode, respectively.}

\item{convertFiles}{Logical, set to \code{TRUE} for non mzML MS files being converted to centroided mzML.
The default is \code{FALSE} for no conversion even if files are present.}

\item{convertFrom}{The name of the file format or vendor format to be found and converted.
Possible entries are: \emph{thermo}, \emph{bruker}, \emph{agilent}, \emph{ab} (from AB Sciex) and \emph{waters}.}

\item{convertToCentroid}{Logical, set to \code{TRUE} to convert profile data to centroided. The default and recommended is \code{TRUE}.}

\item{save}{Logical, set to \code{TRUE} to save the object setup in the \strong{rData} folder.
If \strong{rData} folder does not exist in given \code{projPath}, the folder will be created.}

\item{makeNewProject}{Logical, set to TRUE to create an R project in the given \code{projPath} and open a new R session.}
}
\value{
The output is a simple list, containing: (1) \code{projPath}, the chosen project path; (2) the \code{date}, the given project date; and 
(3) the \code{sampleInfo}, a \code{\link[base]{data.frame}} with the following columns:
(1) \code{filePath}, the path of each mzML file, including any converted MS files;
(2) \code{sample}, the retrieved file name to be used as sample identifier;
(3) \code{group}, the name of each replicate sample group
(e.g. samples with the same name but with different numbering, such as in the case of replicate samples);
(4) \code{blank}, the specified blank group names or the blank group detected by file name (\emph{i.e.} blank, Blank and/or B as file name);
and (5) \code{polarity}, the polarity mode used for each sample, possible entries are \code{positive} and \code{negative};
}
\description{
Project setup by screening a given folder for MS files or
by creating a new folder where the MS files can be added using \code{addFiles}.
Setting \code{createRproject} to \code{TRUE} will create and open an R project in the selected or given folder.
When \code{convertFiles} is \code{TRUE} the function uses \code{\link{mzMLconverter}} to automatically convert specified raw MS files to mzML.
}
\details{
The data.frame \code{sampleInfo} in the output list can always be edited for correction of
sample replicate group names, blank replicate groups and assigned polarity, which is set to \code{positive} by default when not given as argument.
}
\examples{



}
