% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_002_InspectRawData.R
\name{makeTargets}
\alias{makeTargets}
\title{makeTargets}
\usage{
makeTargets(mz, rt, ppm = 20, sec = 60)
}
\arguments{
\item{mz}{A vector with target \emph{m/z} values or
a two columns data.table or data.frame with minimum and maximum \emph{m/z} values.
Alternatively, \emph{m/z} and retention time values can be given as one data.table/data.frame
and the deviations given as \code{ppm} and \code{sec} are used to calculate the ranges.
The same also works for min and max values of \emph{m/z} and retention time targets.
Note that when mass/time ranges are given \code{ppm} and \code{sec} are not used.
An id column with target identifiers can be added with \code{mz} as column.}

\item{rt}{A vector with target retention time values or
a two columns data.table/data.frame with minimum and maximum retention time values.}

\item{ppm}{A numeric vector of length one with the mass deviation, in ppm.}

\item{sec}{A numeric vector of length one with the time deviation, in seconds.}
}
\value{
A data.table with \emph{m/z} and retention time target pairs.
}
\description{
Helper function to build \emph{m/z} and retention time (in seconds)
target pairs for searching within other functions. Each target is composed of an
id and mass (\emph{m/z}) and time (seconds) ranges. When mass is defined without
time, the entire time range is used.
}
