% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_008_uniqueFeatureIdentifiers.R
\name{makeUFI}
\alias{makeUFI}
\title{makeUFI}
\usage{
makeUFI(object, save = FALSE)
}
\arguments{
\item{object}{A \linkS4class{ntsData} object containing annotated features.}

\item{save}{Logical, set to \code{TRUE} to save updated
\linkS4class{ntsData} object in the \strong{rdata} folder.
Note that \code{TRUE} overwrites the existing \linkS4class{ntsData} object.
Optionally, a character string can be given instead of \code{TRUE}
to be used as file name, avoiding overwriting.}
}
\description{
Function to collapse istopes to the respective feature,
building a unique feature identifier (UFI) based on the calculated
neutral monoisotopic mass of the (de)protonated ion.
}
