% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{mapPeaks}
\alias{mapPeaks}
\title{mapPeaks}
\usage{
mapPeaks(object, ...)
}
\arguments{
\item{object}{An object to with EICs for mapping chromatographic peaks.}

\item{...}{Other method dependent parameters.}
}
\value{
A plot with mapped peaks, including time and mass peak space.
}
\description{
mapPeaks
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
