% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_010_filters.R
\name{maxReplicateIntensityDeviation}
\alias{maxReplicateIntensityDeviation}
\title{maxReplicateIntensityDeviation}
\usage{
maxReplicateIntensityDeviation(obj, value = 30)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{value}{A numerical value set at the desired sd percentage maximum.}
}
\description{
maxReplicateIntensityDeviation
}
