% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{metadata}
\alias{metadata}
\title{metadata}
\usage{
metadata(x, ...)
}
\arguments{
\item{x}{Object to collect metadata.}

\item{...}{Other argumments.}
}
\value{
A \link[data.table]{data.table} with metadata in x.
}
\description{
metadata
}
