% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_010_filters.R
\name{minReplicateAbundance}
\alias{minReplicateAbundance}
\title{minReplicateAbundance}
\usage{
minReplicateAbundance(obj, value = 3)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{value}{A numerical value set at the desired minimum representation in replicates.}
}
\description{
minReplicateAbundance
}
