% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_000_S4_parameterClasses.R
\docType{class}
\name{MS2param-class}
\alias{MS2param-class}
\title{MS2param-class}
\value{
An \linkS4class{MS2param} object containing parameters for
extraction of MS2 spectra of given precursor ions in an \linkS4class{ntsData} object.
}
\description{
MS2param-class
}
\section{Slots}{

\describe{
\item{\code{maxMSRtWindow}}{.}

\item{\code{precursorMzWindow}}{.}

\item{\code{clusterMzWindow}}{.}

\item{\code{topMost}}{.}

\item{\code{minIntensityPre}}{.}

\item{\code{minIntensityPost}}{.}
}}

