% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_000_S4_parameterClasses.R
\name{MS2param}
\alias{MS2param}
\title{MS2param}
\usage{
MS2param(
  maxMSRtWindow = 10,
  precursorMzWindow = 1.3,
  clusterMzWindow = 0.003,
  topMost = 50,
  minIntensityPre = 10,
  minIntensityPost = 10
)
}
\arguments{
\item{maxMSRtWindow}{.}

\item{precursorMzWindow}{.}

\item{clusterMzWindow}{.}

\item{topMost}{.}

\item{minIntensityPre}{.}

\item{minIntensityPost}{.}
}
\value{
An \linkS4class{MS2param} object containing parameters for
extraction of MS2 spectra of given precursor ions in an \linkS4class{ntsData} object.
}
\description{
MS2param
}
