% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{MS2s}
\alias{MS2s}
\title{MS2s}
\usage{
MS2s(object, ...)
}
\arguments{
\item{object}{Object to get level 2 fragmentation data (MS/MS or MS2).}
}
\value{
A \link[data.table]{data.table} with MS2 from the object.
}
\description{
MS2s
}
