% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_001_makeSetup.R
\name{mzMLconverter}
\alias{mzMLconverter}
\title{mzMLconverter}
\usage{
mzMLconverter(
  path = setup$projPath,
  convertFrom = "agilent",
  centroidData = TRUE
)
}
\arguments{
\item{path}{The \code{path} where the files to be converted can be found. The function will look into subfolders.}

\item{convertFrom}{The format/vendor of files to be found and converted.
Possible entries are: \emph{thermo}, \emph{bruker}, \emph{agilent}, \emph{ab} (from AB Sciex) and \emph{waters}.}

\item{centroidData}{Logical, set to \code{TRUE} to convert profile data into centroided. The default and recommended is TRUE.}
}
\value{
The converted files will be saved as mzML in the given path.
}
\description{
Converts MS files from different formats to mzML format.
See \pkg{patRoon} \insertCite{Helmus2021}{ntsIUTA} for all the possible formats. 
Optionally and recommended, it centroids the data.
}
\examples{



}
\references{
\insertRef{Helmus2021}{ntsIUTA}
}
