% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_001_supportTools.R
\name{mzrBuilder}
\alias{mzrBuilder}
\title{mzrBuilder}
\usage{
mzrBuilder(mz = NULL, ppm = NULL)
}
\arguments{
\item{mz}{The target \emph{m/z} to calculate range.}

\item{ppm}{The mass deviation in ppm to calculate the \emph{m/z} range.}
}
\description{
mzrBuilder
}
