% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_011_findFragments.R
\docType{class}
\name{ntsFragments-class}
\alias{ntsFragments-class}
\title{ntsFragments}
\value{
An \linkS4class{ntsFragments} object to be added to
the workflows slot of an \linkS4class{ntsData} object.
}
\description{
ntsFragments
}
\section{Slots}{

\describe{
\item{\code{database}}{A \link[data.table]{data.table} with target fragments
for searching in MS2 data of a given feature list.}

\item{\code{settings}}{The parameter settings used for find the fragments.}

\item{\code{data}}{Extra data produced during screening for fragments.}

\item{\code{results}}{A \link[data.table]{data.table}
with summarized results per sample replicate group.}
}}

