% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_011_findFragments.R
\docType{class}
\name{ntsFragmentsData-class}
\alias{ntsFragmentsData-class}
\title{ntsFragmentsData}
\value{
An \linkS4class{ntsFragmentsData} object to be added to
the workflows slot of an \linkS4class{ntsData} object.
}
\description{
ntsFragmentsData
}
\section{Slots}{

\describe{
\item{\code{targets}}{A data frame with target values for finding fragments in MS2 data of a given feature list.}

\item{\code{param}}{The parameters used for find fragments.}

\item{\code{data}}{Extra data produced during screening for fragments.}

\item{\code{results}}{A data.frame with summarized results per sample replicate group.}
}}

