% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_012_monitoring.R
\docType{class}
\name{ntsMonitoringData-class}
\alias{ntsMonitoringData-class}
\alias{exportPlots,ntsMonitoringData,character-method}
\title{ntsMonitoringData}
\usage{
\S4method{exportPlots}{ntsMonitoringData,character}(
  object,
  path = NULL,
  sequences = NULL,
  yUnit = "mz",
  size = NULL,
  ID = NULL,
  interactive = FALSE,
  clusters = NULL,
  ncol = 3
)
}
\arguments{
\item{object}{An \linkS4class{ntsMonitoringData} object.}

\item{path}{A character vector with the path to export the monitoring results.}

\item{sequences}{Other parameters used in the production of process monitoring results.}

\item{yUnit}{The unit to show in the y axis.
Possible values are "mz" and "rt".}

\item{size}{The size for plotting the dots.}

\item{ID}{Optional character vector with features ID to filter the process data.}

\item{interactive}{Logical, set to \code{TRUE} to plot an interactive plot.}

\item{clusters}{The number of clusters to plot.}

\item{ncol}{The number of columns to arrange the clusters grid.}
}
\value{
An \linkS4class{ntsFragmentsData} object to be added to
the workflows slot of an \linkS4class{ntsData} object.
}
\description{
ntsMonitoringData
}
\section{Methods (by generic)}{
\itemize{
\item \code{exportPlots}: Method to export plots.
}}

\section{Slots}{

\describe{
\item{\code{sequences}}{A list of sequences reflecting the treatment/process chain.
The treatments/processes names in each sequence must match sample replicate group names
in an associated \linkS4class{ntsData} object.}

\item{\code{clusters}}{A list of results from the clustering analysis for each sequence.}

\item{\code{categories}}{A list of results from the categorization of each step in each sequence.}

\item{\code{efficiency}}{A list of data.frames with summarized categories results per sequence.}
}}

