% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{ntsParameters-class}
\alias{ntsParameters-class}
\title{ntsParameters}
\value{
A \linkS4class{ntsParameters} object to be used for the basic workflow of \pkg{ntsIUTA}.
}
\description{
A S4 class object to store the parameter settings for
each data processing step of the basic workflow.
Each parameter set is stored as \linkS4class{ntsSettings} object.
See \code{?"ntsSettings-class"}.
}
\section{Slots}{

\describe{
\item{\code{picking}}{Settings for peak picking.}

\item{\code{grouping}}{Settings for grouping and alignment within replicates.}

\item{\code{filling}}{Settings for filling peaks with recursive integration within replicates.}

\item{\code{annotation}}{Settings for annotation of isotopes and adducts.}

\item{\code{fragments}}{Settings for extracting MS2 data.}

\item{\code{unification}}{Settings for unifying features across replicates.}
}}

