% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{ntsSettings-class}
\alias{ntsSettings-class}
\title{ntsSettings}
\value{
A \linkS4class{ntsSettings} object to be used for
the basic workflow of \pkg{ntsIUTA}.
}
\description{
A set of settings for a data processing step of the basic workflow.
The \linkS4class{ntsSettings} object contains the algorithm to be used and the list of respective settings.
}
\section{Slots}{

\describe{
\item{\code{algorithm}}{A character string with the name of the algorithm to be used.}

\item{\code{settings}}{A list of settings dependent on the algorithm used.}
}}

