% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_010_suspectScreening.R
\docType{class}
\name{ntsSuspectData-class}
\alias{ntsSuspectData-class}
\title{ntsSuspectData}
\value{
An \linkS4class{ntsSuspectData} object to be added to
the workflows slot of an \linkS4class{ntsData} object.
}
\description{
ntsSuspectData
}
\section{Slots}{

\describe{
\item{\code{suspectList}}{The \linkS4class{suspectList} object used for suspect screening.}

\item{\code{param}}{The parameters used for suspect screening.}

\item{\code{data}}{The raw \linkS4class{featureGroupsScreening}
for each sample replicate group.}

\item{\code{inSilico}}{When present, the results from the matched in silico
fragmentation experiment.}

\item{\code{results}}{A data.frame with summarized results, including
the identification level according to the guideline.}
}}

