% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_002_InspectRawData.R
\name{openCacheDBScope}
\alias{openCacheDBScope}
\title{openCacheDBScope}
\usage{
openCacheDBScope(
  new = list(),
  file = getCacheFile(),
  .local_envir = parent.frame()
)
}
\description{
Opens the cache database for storing processing data.
The function is loaded from \pkg{patroon}.
}
