% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{paramMS2-class}
\alias{paramMS2-class}
\title{paramMS2-class}
\value{
An \linkS4class{paramMS2} object containing parameters for
extraction of MS2 spectra of given precursor ions in an \linkS4class{ntsData} object.
}
\description{
paramMS2-class
}
\section{Slots}{

\describe{
\item{\code{maxMSRtWindow}}{.}

\item{\code{precursorMzWindow}}{.}

\item{\code{clusterMzWindow}}{.}

\item{\code{topMost}}{.}

\item{\code{minIntensityPre}}{.}

\item{\code{minIntensityPost}}{.}
}}

