% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_methods_002_parameters.R
\name{paramMS2}
\alias{paramMS2}
\title{paramMS2}
\usage{
paramMS2(
  maxMSRtWindow = 10,
  precursorMzWindow = 1.3,
  clusterMzWindow = 0.003,
  topMost = 50,
  minIntensityPre = 10,
  minIntensityPost = 10
)
}
\arguments{
\item{maxMSRtWindow}{.}

\item{precursorMzWindow}{.}

\item{clusterMzWindow}{.}

\item{topMost}{.}

\item{minIntensityPre}{.}

\item{minIntensityPost}{.}
}
\value{
An \linkS4class{paramMS2} object containing parameters for
extraction of MS2 spectra of given precursor ions in an \linkS4class{ntsData} object.
}
\description{
paramMS2
}
