% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{paramSet-class}
\alias{paramSet-class}
\title{paramSet}
\value{
A \linkS4class{paramSet} object to be used for
the basic workflow of \pkg{ntsIUTA}.
}
\description{
A set of parameters for a data processing step of the basic workflow.
The \linkS4class{paramSet} object contains the algorithm to be used and the list of respective parameters.
}
\section{Slots}{

\describe{
\item{\code{algorithm}}{A character string with the name of the algorithm to be used.}

\item{\code{param}}{A list of parameters dependent on the algorithm used.}
}}

