% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{path}
\alias{path}
\title{path}
\usage{
path(object, ...)
}
\arguments{
\item{object}{Object to change or get the information.}

\item{...}{Other arguments.}
}
\value{
Get the path of the object.
}
\description{
path
}
