% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_007_peakAnnotation.R
\name{peakAnnotation}
\alias{peakAnnotation}
\title{peakAnnotation}
\usage{
peakAnnotation(object = NULL, algorithm = NULL, settings = NULL, save = FALSE)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object with peaks/features.}

\item{algorithm}{The algorithm for annotation.}

\item{settings}{The respective parameter settings.}

\item{save}{Logical, set to \code{TRUE} to save updated
\linkS4class{ntsData} object in the \strong{rdata} folder.
Note that \code{TRUE} overwrites the existing \linkS4class{ntsData} object.
Optionally, a character string can be given instead of \code{TRUE}
to be used as file name, avoiding overwriting.}

\item{excludeBlanks}{Logical, set to \code{TRUE} for excluding
blank replicates from annotation. Blank intensities are not considered.}
}
\value{
An \linkS4class{ntsData} object with annotated peaks/features.
}
\description{
Annotation of isotopic and adduct peaks.
}
