% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_005_peakFilling.R
\name{peakFilling}
\alias{peakFilling}
\title{peakFilling}
\usage{
peakFilling(object, algorithm = NULL, settings = NULL, save = TRUE)
}
\arguments{
\item{XCMSfeatures}{An \linkS4class{XCMSnExp} object containing features.}

\item{paramFill}{An object of class \code{FillChromPeaksParam} or \code{ChromPeakAreaParam}
containing the parameters to apply the recursive integration.
See \code{?\link[xcms]{fillChromPeaks}} for more information.}
}
\value{
An \linkS4class{XCMSnExp} object including filled missing peaks.
}
\description{
Recursive integration for samples with missing peaks
using the function \code{\link[xcms]{fillChromPeaks}}
from the \pkg{xcms} package.
}
\references{
\insertRef{xcms1}{ntsIUTA}
\insertRef{xcms2}{ntsIUTA}
\insertRef{xcms3}{ntsIUTA}
}
