% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_004_peakGrouping.R
\name{peakGrouping}
\alias{peakGrouping}
\title{peakGrouping}
\usage{
peakGrouping(
  object = NULL,
  algorithm = NA_character_,
  settings = NULL,
  simplify = TRUE,
  save = FALSE
)
}
\arguments{
\item{object}{A \linkS4class{ntsData} object containing peaks.}

\item{algorithm}{The algorithm to use for peak alignment and grouping.}

\item{settings}{The respective parameter settings.
See \link[patRoon]{groupFeatures} for more information.}

\item{simplify}{Logical, set to \code{TRUE} to convert the patRoon object
to a \linkS4class{featureGroupsSIRIUS} object which is simpler and faster for subsetting.}

\item{save}{Logical, set to \code{TRUE} to save updated
\linkS4class{ntsData} object in the \strong{rdata} folder.
Note that \code{TRUE} overwrites the existing \linkS4class{ntsData} object.
Optionally, a character string can be given instead of \code{TRUE}
to be used as file name, avoiding overwriting.}
}
\value{
An \linkS4class{ntsData} object
containing peaks aligned and grouped as features across samples.

 @references
\insertRef{Helmus2021}{ntsIUTA}
}
\description{
Grouping and alignment of peaks across samples.
The peak grouping uses the \pkg{patRoon} package and
the following algorithms are possible: "xcms3", "xcms", "openms".
See ?\pkg{patRoon} for further information.
}
