% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_003_peakPicking.R
\name{peakPicking}
\alias{peakPicking}
\title{peakPicking}
\usage{
peakPicking(
  object = NULL,
  algorithm = NA_character_,
  settings = NULL,
  save = FALSE
)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object.}

\item{algorithm}{A character string with the algorithm to use for peak picking.}

\item{settings}{List of parameter settings for the specified algorithm.
See \link[patRoon]{findFeatures} for more information.}

\item{save}{Logical, set to \code{TRUE} to save updated
\linkS4class{ntsData} object in the \strong{rdata} folder.
Note that \code{TRUE} overwrites the existing \linkS4class{ntsData} object.
Optionally, a character string can be given instead of \code{TRUE}
to be used as file name, avoiding overwriting.}
}
\value{
An \linkS4class{ntsData} object with peaks per sample and an updated
\linkS4class{features} object added to the slot \code{pat}.
}
\description{
Finds chromatographic peaks from centroided MS data in the
samples of a given \linkS4class{ntsData} object.
The peak picking uses the \pkg{patRoon} package and
the following algorithms are available:
"xcms3", "xcms", "openms", "envipick", "sirius", "kpic2", "safd".
The parameters depend on the algorithm chosen.
See ?\pkg{patRoon} for further information.
}
\references{
\insertRef{Helmus2021}{ntsIUTA}
}
