% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{peakPickingParameters}
\alias{peakPickingParameters}
\title{peakPickingParameters}
\usage{
peakPickingParameters(object, algorithm, param)
}
\arguments{
\item{object}{An object to add or get peak picking parameters.}

\item{...}{Other method dependent parameters.}
}
\value{
A list of parameters or the object with added parameters.
}
\description{
peakPickingParameters
}
