% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{peaks}
\alias{peaks}
\title{peaks}
\usage{
peaks(object, ...)
}
\arguments{
\item{object}{An object to extract chromatographic peaks.}

\item{...}{Other method dependent parameters.}
}
\value{
A data.frame with peak details.
}
\description{
peaks
}
