% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{pickingParameters}
\alias{pickingParameters}
\title{pickingParameters}
\usage{
pickingParameters(object, algorithm, settings)
}
\arguments{
\item{object}{An object to add or get peak picking parameter settings.}

\item{algorithm}{Algorithm for the parameters.}

\item{settings}{List of respective settings.}
}
\value{
A list of parameter settings or the object with added parameter settings.
}
\description{
pickingParameters
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
