% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_006_alignment.R
\name{plotAlignment}
\alias{plotAlignment}
\title{plotAlignment}
\usage{
plotAlignment(object, samples = NULL, interactive = TRUE)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object with adjusted retention time.}

\item{samples}{A numeric or character vector with the sample to plot the
retention time adjustment.}

\item{interactive}{Logical, set to \code{TRUE} for plotting with \pkg{plotly} package.}
}
\value{
A plot with the retention time alignment differences for each sample.
}
\description{
Plots the results from the retention time alignment across samples.
}
