% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_008_checkInternalStandards.R
\name{plotCheckIS}
\alias{plotCheckIS}
\title{plotCheckIS}
\usage{
plotCheckIS(obj, rtWindow = NULL, ppm = NULL)
}
\arguments{
\item{obj}{obj An \linkS4class{ntsData} object.}

\item{rtWindow}{The time window to show the time deviation.}

\item{ppm}{The ppm window to show the mass deviation.}
}
\value{
A plot of the IS check results.
}
\description{
plotCheckIS
}
