% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{plotComponents}
\alias{plotComponents}
\title{plotComponents}
\usage{
plotComponents(object, ...)
}
\arguments{
\item{object}{An object to extract components.}

\item{...}{Other method dependent parameters.}
}
\value{
A plot of components.
}
\description{
plotComponents
}
