% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_007_components.R
\name{plotComponentsInteractive}
\alias{plotComponentsInteractive}
\title{plotComponentsInteractive}
\usage{
plotComponentsInteractive(object, comps, colorBy = "isotopes")
}
\value{
A dotted plot with unified features.
}
\description{
Function to plot features unified by neutral masss.
The unified features include the monoisotpic trace
and other annotated isotopic and adduct traces.
}
