% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_007_components.R
\name{plotComponentsOld}
\alias{plotComponentsOld}
\title{plotComponentsOld}
\usage{
plotComponentsOld(
  obj = NULL,
  samples = NULL,
  ID = NULL,
  mz = NULL,
  ppm = 5,
  rt = NULL,
  rtWindow = 1,
  rtUnit = "min",
  comp = NULL,
  entireComponents = TRUE,
  onlyAnnotated = FALSE,
  onlyRelated = TRUE,
  log = TRUE,
  colorBy = c("groups")
)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object containing annotated features.}

\item{samples}{The samples (name or index) to extract the features.
The default is \code{NULL} which considers all existing sample replicate groups,
excluding any assigned blank replicate groups as listed in the \code{obj}.
Note that the features are taken from sample replicate groups, meaning that
features will be extracted from sample replicate groups
that contain the specified samples.}

\item{ID}{The identifier/s of selected features. When specified,
it overwrites any given \code{mz} or \code{comp} values.}

\item{mz}{The \emph{m/z} to find features. Not used if \code{ID} is specified.
Can be a vector of length 2, defining the mass range to find features.}

\item{ppm}{The expected mass deviation (in ppm) to search
for features of a given \code{mz}.}

\item{rt}{The expected retention time of the \emph{m/z} of interest,
only used if \code{ID} is not specified.}

\item{rtWindow}{The expected retention time deviation for searching.
Can be a vector of length 2, giving the time range to find features.}

\item{rtUnit}{The time unit used.
Possible values are \code{sec} and \code{min}. Default is \code{sec}.}

\item{comp}{The component number/s to extract features.
Only used if both \code{ID} and \code{mz} are \code{NULL}.}

\item{entireComponents}{Logical, set to \code{TRUE} (The default) to give the
all the features in the components represented by the selected features.}

\item{onlyAnnotated}{Logical, set to \code{TRUE} to return only annotated features.}

\item{onlyRelated}{Logical, set to \code{TRUE} to return only features that are related,
meaning features annotated with the same molecular ion.}

\item{log}{Logical, set to \code{TRUE} (the default) to plot the log of the peak heights.}

\item{colorBy}{Possible values are \code{samples} or \code{groups} (the default)
for plotting by samples or sample replicate groups, respectively. Note that for
\code{groups} the height of the average peaks is averaged.}
}
\value{
A spectrum plot of annotation details for
the selected/found features.
}
\description{
Plots annotation details for features selected by \code{ID},
\emph{m/z} and retention time or component number (\code{comp}) of all
or selected sample replicate groups in an \linkS4class{ntsData} object.
}
\note{
If all (\code{ID}, \code{mz} and \code{comp}) are \code{NULL},
the returned \code{data.frame} will contain all the features
for each sample replicate group. Additionally, the three logical arguments
are applied in the follwoing order: (1) \code{entireComponents},
(2) \code{onlyAnnotated} and (3) \code{onlyRelated}.
}
