% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotComponentSpectrum.R
\name{plotComponentSpectrum}
\alias{plotComponentSpectrum}
\title{plotComponentSpectrum}
\usage{
plotComponentSpectrum(
  xA = featComp,
  replicateGroups = NULL,
  features = NULL,
  featData = featData,
  mz = NULL,
  ppm = 5,
  mzWindowPlot = NULL,
  rt = NULL,
  rtWindow = 1,
  rtUnit = "min",
  comp = NULL,
  onlyAnnotated = FALSE,
  onlyRelated = TRUE,
  intval = "maxo",
  log = TRUE
)
}
\arguments{
\item{xA}{The list of \linkS4class{xsAnnotate} objects obtained via \code{\link{makeFeatureComponents}}.}

\item{replicateGroups}{The replicate groups name or number to filter the \code{xA} list,
which is named according to the experimental replicate groups as defined in \code{\link{setupProject}}.}

\item{features}{Feature identifier/s.}

\item{featData}{The feature data to get the feature identifiers.
Can be a \linkS4class{XCMSnExp} or \linkS4class{featureGroups} object from \pkg{xcms} or \pkg{patRoon}, respectively.}

\item{mz}{The \emph{m/z} of interest. If \code{features} are specified \code{mz is not used}.}

\item{ppm}{The expected mass deviation to search for features of a given \code{mz}.}

\item{mzWindowPlot}{The \emph{m/z} range for the plot.}

\item{rt}{The expected retention time of the \emph{m/z} of interest, only used if \code{features} are not speficied.}

\item{rtWindow}{The expected retention time deviation for searching.}

\item{rtUnit}{The time unit used. Default is minutes.}

\item{comp}{The component numbers to plot. Only used if both \code{features} and \code{mz} are \code{NULL} (i.e. not specified).}

\item{onlyAnnotated}{Plots only annotated features.}

\item{onlyRelated}{Plots only relevant features (i.e. isotopes and addcuts) of the specified features.}

\item{intval}{The intensity value type. Default is "maxo", corresponding to the height.}

\item{log}{Logical, set to \code{TRUE} (the default) to plot the intensities in a log scale.}
}
\value{
A spectrum plot of given components or features.
}
\description{
Plots the spectra for given features, \emph{m/z} and rt or components from a list of \linkS4class{xsAnnotate} objects.
}
\examples{



}
