% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotCheckReplicates.R
\name{plotCorReplicates,ntsData-method}
\alias{plotCorReplicates,ntsData-method}
\title{plotCorReplicates}
\usage{
\S4method{plotCorReplicates}{ntsData}(obj, binsize = 2)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object with more than one file.}

\item{binsize}{The bin size to apply to the raw data. The default is 2.}
}
\value{
A \emph{heatmap} plot to evaluate the correlation.
}
\description{
Evaluates the quality of replicate sample groups using a pearson
correlation of binned raw data.
}
