% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{plotEICs}
\alias{plotEICs}
\title{plotEICs}
\usage{
plotEICs(object, ...)
}
\arguments{
\item{object}{Object to extract extract ion chromatograms (EICs).}
}
\value{
A plot with EICs for defined samples in the object.
}
\description{
plotEICs
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
