% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotFeaturePeaks.R
\name{plotFeaturePeaks}
\alias{plotFeaturePeaks}
\title{plotFeaturePeaks}
\usage{
plotFeaturePeaks(
  x = featData,
  fileIndex = NULL,
  features = NULL,
  mz = NULL,
  ppm = 5,
  rt = NULL,
  rtWindow = 1,
  rtUnit = "min",
  plotBy = "samples",
  names = NULL
)
}
\arguments{
\item{x}{An \linkS4class{XCMSnExp} or \linkS4class{featureGroups} object with one or more files and grouped peaks.}

\item{fileIndex}{The index of the file/s to extract the centroids or profile data.}

\item{features}{The identifier of the features of interest.}

\item{mz}{Optional target \emph{m/z} to find features using the mass deviation specified by the \code{ppm} parameter.}

\item{ppm}{The mass deviation to extract the features when \code{mz} is specified.}

\item{rt}{The retention time in minutes or seconds, depending on the defined \code{rtUnit}, see below. Only used when \code{mz} is specified.}

\item{rtWindow}{The time deviation to collect features. The time unit is the defined by \code{rtUnit}.
A time interval can be given with a length 2 vector, defining the minimum and maximum retention time.}

\item{rtUnit}{Possible entries are \code{min} or \code{sec}. The default is \code{min}.}

\item{plotBy}{Grouping for legend. Possible entries are: \code{samples}, \code{features} or \code{groups}.}

\item{names}{A character string with names for each feature given in \code{features}.
Note that length should match between \code{names} and \code{features}.}
}
\value{
A double plot with peaks chromatograms on the top part and feature peak groups below.
}
\description{
Plots feature information for a given \linkS4class{XCMSnExp} object containing grouped peaks across samples.
}
\examples{



}
