% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{plotMS2s}
\alias{plotMS2s}
\title{plotMS2s}
\usage{
plotMS2s(object, ...)
}
\arguments{
\item{object}{Object to get level 2 fragmentation data (MS/MS or MS2).}
}
\value{
A plot with MS2 from the object.
}
\description{
plotMS2s
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
