% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotPeaks.R
\name{plotPeaks}
\alias{plotPeaks}
\title{plotPeaks}
\usage{
plotPeaks(
  obj,
  samples = NULL,
  ID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  rtWindow = NULL,
  rtUnit = "sec",
  colorBy = "samples"
)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{samples}{The index or name of the sample/s.
The default is \code{NULL} and all samples are used.}

\item{ID}{The identifier of the peaks of interest.
When not \code{NULL}, overwrites any given \code{mz} and \code{rt} value.}

\item{mz}{Optional target \emph{m/z} to find peaks using
the mass deviation specified by the \code{ppm} parameter.}

\item{ppm}{The mass deviation to extract the peaks
when \code{mz} is specified.}

\item{rt}{The retention time in minutes or seconds,
depending on the defined \code{rtUnit}, see below.
Only used when \code{mz} is specified.}

\item{rtWindow}{The time deviation to collect peaks.
The time unit is the defined by \code{rtUnit}.
A time interval can be given with a length 2 vector,
defining the minimum and maximum retention time.}

\item{rtUnit}{Possible entries are \code{min} or \code{sec}.
The default is \code{min}.}

\item{colorBy}{Possible values are \code{"peaks"} (the default),
\code{"samples"} or \code{sampleGroups},
for colouring by peaks, samples or sample replicate groups, respectively.}
}
\value{
A peak/s map plot produced through \pkg{plotly}.
}
\description{
Method for plotting peaks from a \linkS4class{ntsData} object.
}
