% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotMonitoring.R
\name{plotProcessCategories}
\alias{plotProcessCategories}
\title{plotProcessCategories}
\usage{
plotProcessCategories(
  obj,
  sequences = NULL,
  ID = NULL,
  yUnit = "mz",
  size = NULL,
  interactive = TRUE
)
}
\arguments{
\item{obj}{An \linkS4class{ntsMonitoringData} object with monitoring categories to plot.}

\item{sequences}{Optional vector with sequence names from the \linkS4class{ntsMonitoringData} object to plot.
When \code{NULL}, all sequences are plotted.}

\item{ID}{Optional character vector with features ID to filter the process data.}

\item{yUnit}{The unit to use for the y axis. Possible values are "mz" or "rt.}

\item{size}{The size for ploting the dots.}

\item{interactive}{Logical, set to \code{TRUE} to plot an interactive plot.}
}
\value{
A list with plots for each given sequence in the \linkS4class{ntsMonitoringData} object.
}
\description{
plotProcessCategories
}
