% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotMonitoring.R
\name{plotProcessClusters}
\alias{plotProcessClusters}
\title{plotProcessClusters}
\usage{
plotProcessClusters(
  obj,
  sequences = NULL,
  clusters = NULL,
  ID = NULL,
  interactive = FALSE,
  ncol = 3
)
}
\arguments{
\item{obj}{An \linkS4class{ntsMonitoringData} object with cluster analysis for process evaluation.}

\item{sequences}{Optional vector with sequence names from the \linkS4class{ntsMonitoringData} object to plot.
When \code{NULL}, all sequences are plotted.}

\item{clusters}{The number of clusters to plot.}

\item{ID}{The ID of features to plot.}

\item{interactive}{Logical, set to \code{TRUE} for using plotly instead of ggplot2.}

\item{ncol}{The number of columns to arrange the clusters grid.}
}
\value{
A grid plot list with the intensity trends of features in each cluster for each sequence.
}
\description{
plotProcessClusters
}
