% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotTargetCentroids.R
\name{plotTargetCentroids}
\alias{plotTargetCentroids}
\title{plotTargetCentroids}
\usage{
plotTargetCentroids(
  x = rawData,
  fileIndex = 1,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  rtWindow = NULL,
  rtUnit = "min",
  title = NULL,
  plotTargetMark = TRUE,
  ...
)
}
\arguments{
\item{x}{An \linkS4class{OnDiskMSnExp} object with one or more files.}

\item{fileIndex}{The index of the file/s to extract the centroids or profile data.}

\item{mz}{The target \emph{m/z}. Note that \code{m/z} is not the expected monoisotopic mass but the expected adduct (\emph{i.e.} \code{[M+H]+}).}

\item{ppm}{The mass deviation to extract centroids in \code{ppm}.}

\item{rt}{The expected retention time in minutes or seconds, depending on the defined \code{rtUnit}, see below.}

\item{rtWindow}{The time deviation to collect centroids or profile data. The time unit is the defined by \code{rtUnit}.
It is recommended a minimum of 1 minute.}

\item{rtUnit}{Possible entries are \code{min} or \code{sec}. The default is \code{sec}.}

\item{title}{Optional title for the plot/s.}

\item{plotTargetMark}{Logical, set to TRUE, the defaulft, to plot a target mark with +/- 5 ppm and +/- 10 seconds deviation.}

\item{...}{Other arguments to be paced to the ploting function.}
}
\value{
A plot of the centroids for the requested target.
}
\description{
Plot centroids or profile data for a target compound using expected \emph{m/z}, retention time and respective deviations.
}
\examples{



}
