% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{plotTICs}
\alias{plotTICs}
\title{plotTICs}
\usage{
plotTICs(object, ...)
}
\arguments{
\item{object}{Object to extract total ion chromatograms (TICs).}
}
\value{
A plot with TICs from the object.
}
\description{
plotTICs
}
