% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_002_XICs.R
\name{plotXIC_base}
\alias{plotXIC_base}
\title{plotXIC_base}
\usage{
plotXIC_base(
  xic,
  plotTargetMark = TRUE,
  ppmMark = 5,
  secMark = 10,
  numberRows = 1
)
}
\arguments{
\item{plotTargetMark}{Logical, set to \code{TRUE} (the defaulft) to plot a
target mark.}

\item{ppmMark}{A numeric vector of length one to define the mass deviation, in ppm,
of the target mark. The default is 5 ppm.}

\item{secMark}{A numeric vector of length one to define the time deviation, in seconds,
of the target mark. The default is 10 ppm.}

\item{numberRows}{A numeric vector of length one to define
the number of rows to grid the plots.
Note that each target is always plotted in one row for all selected samples.}

\item{object}{A \link[data.table]{data.table} from the method \link{XICs}.}
}
\value{
An iterative plot of the centroids for the requested
\emph{m/z} and retention time pairs.
}
\description{
Plot centroids or profile data for a target compound
using expected \emph{m/z} and retention time pairs, including deviations.
}
