% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{plotXICs}
\alias{plotXICs}
\title{plotXICs}
\usage{
plotXICs(object, ...)
}
\arguments{
\item{object}{Object to plot three dimentional (\emph{m/z}, time and intensity)
extracted ion chromatograms (XICs).}
}
\value{
A plot with XICs from the object.
}
\description{
plotXICs
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
