% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{polarity<-}
\alias{polarity<-}
\title{polarity<-}
\usage{
polarity(object) <- value
}
\arguments{
\item{object}{Object to assign polarity modes.}

\item{value}{A character vector with polarities mode to assign to the object.}
}
\value{
An object updated with polarity modes.
}
\description{
polarity<-
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
