% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{polarity}
\alias{polarity}
\title{polarity}
\usage{
polarity(object, ...)
}
\arguments{
\item{object}{Object to collect polarities.}

\item{...}{Other arguments.}
}
\value{
A vector with polarity modes for entries in the object.
}
\description{
polarity
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
