% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_012_monitoring.R
\name{processMonitoring}
\alias{processMonitoring}
\title{processMonitoring}
\usage{
processMonitoring(
  obj,
  sequences = NULL,
  title = NULL,
  constantLevel = 5000,
  doClustering = FALSE,
  minClust = 10,
  sizeClus = 2:100,
  exportResults = FALSE
)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object with features for process monitoring.}

\item{sequences}{A list of sequences reflecting the treatment/process chain.
The treatments/processes names in each sequence must match sample replicate group names
in an associated \linkS4class{ntsData} object.}

\item{title}{Optional title for the workflow object entry in the \linkS4class{ntsData} object.}

\item{constantLevel}{The level, in counts, for assuming constant features through steps.}

\item{doClustering}{Logical, set to \code{TRUE} to perform clustering analysis for each sequence.}

\item{minClust}{The minimum number of clusters.}

\item{sizeClus}{The size range for clustering. The default is \code{2:100}.}

\item{exportResults}{Logical, set to \code{TRUE} to export
the standard monitoring results to the results forder.}
}
\value{
An \linkS4class{ntsMonitoringData} object added to the workflows slot
of the \linkS4class{ntsData} object.
}
\description{
processMonitoring
}
