% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{projectInfo}
\alias{projectInfo}
\title{projectInfo}
\usage{
projectInfo(object, ...)
}
\arguments{
\item{object}{Object to change or get the information.}
}
\value{
Get or change title, description and/or date of the a project.
}
\description{
projectInfo
}
