% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{QC<-}
\alias{QC<-}
\title{QC<-}
\usage{
QC(object, ...) <- value
}
\arguments{
\item{object}{Object to collect the samples to be used as QC.}

\item{value}{A character vector the names of the sample replicate group/s
to be used as QC.}

\item{...}{Other argumments.}
}
\value{
Ac object with experimental samples assigned to QC.
}
\description{
QC<-
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
