% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{QC}
\alias{QC}
\title{QC}
\usage{
QC(object)
}
\arguments{
\item{object}{Object to collect the samples used for QC.}
}
\value{
A vector with sample names used for QC.
}
\description{
QC
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
