% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_011_qualityControl.R
\name{qc2ntsData}
\alias{qc2ntsData}
\title{qc2ntsData}
\usage{
qc2ntsData(obj)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object containing QC results.}
}
\value{
An \linkS4class{ntsData} object with QC data added to the main data slots.
}
\description{
Converts the QC data in an \linkS4class{ntsData} object to a new
\linkS4class{ntsData} object.
}
