% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{qcData-class}
\alias{qcData-class}
\title{qcData}
\value{
A \linkS4class{qcData} object to store the QC parameters, data and results
from an \linkS4class{ntsData} object.
}
\description{
qcData
}
\section{Slots}{

\describe{
\item{\code{samples}}{A data.frame listing the QC samples.}

\item{\code{targets}}{A \linkS4class{suspectList} object with the QC target compounds.}

\item{\code{rtWindow}}{The retention time window, in seconds, to look for target compounds.}

\item{\code{ppm}}{The mass window, in ppm, to look for target compounds.}

\item{\code{MSnExp}}{An \linkS4class{OnDiskMSnExp} object with QC raw data.}

\item{\code{patdata}}{A \linkS4class{features} or \linkS4class{featureGroups}
object derived from the basic NTS workflow (peak picking, alignment and grouping).}

\item{\code{peaks}}{A data.frame containing the peak information for each QC sample.}

\item{\code{features}}{A data.frame with the features information.}

\item{\code{annotation}}{A list with the annotation for the QC samples.}

\item{\code{results}}{A data frame with the QC results summary.}
}}

