% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_001_makeSetup.R
\name{removeDuplicateNames}
\alias{removeDuplicateNames}
\title{removeDupliceNames}
\usage{
removeDuplicateNames(orgInfo, newInfo)
}
\arguments{
\item{orgInfo}{Original dataframe where additional entries are to be added}

\item{newInfo}{Dataframe containing the new samples to be added to the original data.}
}
\value{
Returns a dataframe containing the new entries which changed sample names in case of duplicates.
}
\description{
Removes duplicate sample names to avoid problems during further analysis.
}
