% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_007_filterFeatures.R
\name{removeFilteredFeatures}
\alias{removeFilteredFeatures}
\title{removeFilteredFeatures}
\usage{
removeFilteredFeatures(obj)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}
}
\value{
An \linkS4class{ntsData} object with filtered features moved to the removed slot.
}
\description{
removeFilteredFeatures
}
