% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_001_setupProject.R
\name{removeMetadata}
\alias{removeMetadata}
\title{removeMetadata}
\usage{
removeMetadata(object, varname = NULL)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object.}

\item{varname}{A character vector with column metadata name/s
to be removed from the object. If \code{NULL}, the default,
all metadata is removed.}
}
\value{
An \linkS4class{ntsData} with specified metadata removed for the slot metadata.
}
\description{
Removes metadata in an \linkS4class{ntsData} object.
}
