% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{replicates<-}
\alias{replicates<-}
\title{replicates<-}
\usage{
replicates(object) <- value
}
\arguments{
\item{object}{Object to assign sample replicate names.}

\item{value}{A character vector with one or more sample replicate names.}
}
\value{
An object updated with sample replicate names.
}
\description{
replicates<-
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
