% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{replicates}
\alias{replicates}
\title{replicates}
\usage{
replicates(object)
}
\arguments{
\item{object}{Object to collect sample replicate names.}
}
\value{
A vector with sample replicate names.
}
\description{
replicates
}
