% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_010_filters.R
\name{restoreFilteredFeatures}
\alias{restoreFilteredFeatures}
\title{restoreFilteredFeatures}
\usage{
restoreFilteredFeatures(obj)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}
}
\value{
An \linkS4class{ntsData} object with filtered features restored to the features slot.
}
\description{
restoreFilteredFeatures
}
