% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_000_S4_generics.R
\name{sampleGroups<-}
\alias{sampleGroups<-}
\title{sampleGroups<-}
\usage{
sampleGroups(object) <- value
}
\arguments{
\item{object}{Object to assign sample replicate groups.}

\item{value}{A character vector with one or more sample replicate group names.}
}
\value{
An object updated with sample replicate group names.
}
\description{
sampleGroups<-
}
