% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{sampleGroups}
\alias{sampleGroups}
\title{sampleGroups}
\usage{
sampleGroups(object)
}
\arguments{
\item{object}{Object to collect sample replicate groups.}
}
\value{
A vector with sample replicate group names.
}
\description{
sampleGroups
}
