% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{samples}
\alias{samples}
\title{samples}
\usage{
samples(object)
}
\arguments{
\item{object}{Object to collect sample names.}
}
\value{
A vector with sample names.
}
\description{
samples
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
