% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{samplesTable}
\alias{samplesTable}
\title{samplesTable}
\usage{
samplesTable(object)
}
\arguments{
\item{object}{Object to collect samples table.}
}
\value{
A data table with sample information.
}
\description{
samplesTable
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
