% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{saveParameters}
\alias{saveParameters}
\title{saveParameters}
\usage{
saveParameters(object, ...)
}
\arguments{
\item{object}{An object with parameter settings to save.}

\item{...}{Other arguments.}
}
\value{
The parameters are saved in a location to define.
}
\description{
saveParameters
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
