% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_010_suspectScreening.R
\name{screenSuspectsFromCSV}
\alias{screenSuspectsFromCSV}
\title{screenSuspectsFromCSV}
\usage{
screenSuspectsFromCSV(
  patData = patData,
  polarity = "positive",
  ppm = 10,
  rtWindow = 15,
  DB = utils::read.csv(base::file.choose()),
  withMS2 = FALSE,
  listMS2 = NULL
)
}
\arguments{
\item{patData}{A \linkS4class{featureGroups} object with the feature groups of interest}

\item{polarity}{The polatiry of the MS files in \code{patData}. Possible values are 'positive' or 'negative'.}

\item{ppm}{The allowed mass deviation, in ppm.}

\item{rtWindow}{The allowed retention time window, in seconds, when 'rt' is given in the screening list (\code{DB}).}

\item{DB}{A screeningList.}

\item{withMS2}{Logical, set to \code{TRUE} for using MS2 for confirming suspects.}

\item{listMS2}{An MS2 list, if not given it will be calculated.}
}
\value{

}
\description{
screenSuspectsFromCSV
}
\examples{



}
\references{
\insertRef{Helmus2021}{ntsIUTA}
}
