% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_010_filters.R
\name{snRatio}
\alias{snRatio}
\title{snRatio}
\usage{
snRatio(obj, value = 10)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{snRatio}{A numerical value set at the desired signal-to-noise ratio for features.}
}
\description{
snRatio
}
