% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_000_S4_suspectList.R
\docType{class}
\name{suspectList-class}
\alias{suspectList-class}
\title{suspectList}
\value{
An \linkS4class{suspectList} object to be used within screening
workflows of \pkg{ntsIUTA}.
}
\description{
suspectList
}
\section{Slots}{

\describe{
\item{\code{data}}{The data.frame listing all the suspects of interest.}

\item{\code{length}}{The number of compounds in the suspect list.}

\item{\code{rtUnit}}{The general time unit used for the retention time.
The default and recommended is "sec".}

\item{\code{comment}}{Optional comment for the suspect list.}
}}

