% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_010_suspectScreening.R
\name{suspectScreening}
\alias{suspectScreening}
\title{suspectScreening}
\usage{
suspectScreening(
  obj,
  samples = NULL,
  ID = NULL,
  title = NULL,
  suspectList = NULL,
  ppm = 5,
  rtWindow = 30,
  adduct = NULL,
  excludeBlanks = TRUE,
  withMS2 = TRUE,
  param = paramMS2()
)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object with features
to preform suspect screening.}

\item{samples}{A character vector with the name of the
samples to perform the suspect screening.}

\item{ID}{A character vector with the ID of the features to perform suspect screening.}

\item{title}{An optional string (without spaces) to be used as identifier for the
suspect screening entry in the \code{workflows} slot of the \linkS4class{ntsData} object.}

\item{suspectList}{A \linkS4class{suspectList} object containing suspect compounds.
See \link{getSuspectListTemplate} for more details about the information required.}

\item{ppm}{The mass deviation, in ppm, to screen for the suspects.
The default is 5 ppm.}

\item{rtWindow}{The retention time deviation, in seconds,
to screen for the QC target standards. The default is 30 seconds.}

\item{adduct}{The adduct class for screening suspects. The default is \code{NULL},
which uses the polarity of the \code{obj} or the \code{adduct} class defined
in the suspects data frame. See details for more information.}

\item{excludeBlanks}{Logical, set to \code{TRUE} to ignore replicate groups
assigned as blanks in the \code{samples} slot of the \code{obj}.}

\item{withMS2}{Logical, set to \code{TRUE} for using confirmation via MS2.}

\item{param}{A paramMS2 object with parameters for MS2 extraction.}
}
\value{
An \linkS4class{ntsData} object with screening results
added to the workflows slot.
}
\description{
Method to perform suspect screening from a given
list of candidates in a csv file. The method uses the S4 method
\code{screenSuspects} from \pkg{patRoon}.
}
\references{
\insertRef{Helmus2021}{ntsIUTA}
}
