% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{TICs}
\alias{TICs}
\title{TICs}
\usage{
TICs(object, ...)
}
\arguments{
\item{object}{Object to extract total ion chromatograms (TICs).}
}
\value{
A \link[data.table]{data.table} with TICs from the object.
}
\description{
TICs
}
