% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{XICs}
\alias{XICs}
\title{XICs}
\usage{
XICs(object, ...)
}
\arguments{
\item{object}{Object to get three dimentional (\emph{m/z}, time and intensity)
extracted ion chromatograms (XICs).}
}
\value{
A \link[data.table]{data.table} with XICs from the object.
}
\description{
XICs
}
