% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_001_makeSetup.R
\name{addFiles}
\alias{addFiles}
\title{addFiles}
\usage{
addFiles(
  newFiles = utils::choose.files(),
  sampleInfo = data.frame(filePath = character(), sample = character(), group =
    character(), blank = character(), polarity = character(), date = character(), wdir =
    character()),
  copyFiles = FALSE,
  projPath = getwd(),
  date = Sys.Date(),
  groups = NULL,
  polarity = "positive"
)
}
\arguments{
\item{newFiles}{A list of paths for selected mzML or mzXML files to be added to the project. The default is a UI to choose the files.}

\item{sampleInfo}{A sampleInfo \code{data.frame} as created by \code{\link{setupProject}} to add the selected files.}

\item{copyFiles}{Logical. Set to \code{TRUE} to copied the selected files to the defined \code{path}.
\code{FALSE} will add the original location of the files to the sampleInfo.
Note that if duplicate names exist, the copied files are renamed or when \code{copyFiles} is \code{FALSE}, the original files will be renamed.}

\item{projPath}{When \code{copyFiles} is \code{TRUE}, the project \code{path} to copy the files.
The default is the working directory of the project given by \code{base::getwd()}.}

\item{date}{The project date. The default is the system date generated by \code{\link[base]{Sys.time}}.}

\item{groups}{A vector with the identifier/name of each sample replicate group. If \code{NULL},
a grouping tentative will be made using the name of the MS files found in the given \code{path}.}

\item{polarity}{A vector specifying the polarity mode used in each MS file.
Possible values are \code{positive} or \code{negative} for positive and negative mode, respectively.}
}
\value{
Returns an updated sampleInfo \code{dataframe} object.
}
\description{
Adds mzML or mzXML files to the sampleInfo \code{data.frame} object of the project
as defined by the function \code{\link{setupProject}}.
}
\details{
The groups, blanks and polarity can be altered afterwards directly on the sampleInfo \code{dataframe} object.
}
\examples{

}
