% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_001_setupProject.R
\name{addFiles}
\alias{addFiles}
\title{addFiles}
\usage{
addFiles(
  files = utils::choose.files(),
  object = ntsData,
  copy = FALSE,
  replicates = NULL,
  polarity = "positive",
  method = NULL
)
}
\arguments{
\item{files}{A list of paths for selected mzML and/or mzXML files.
The default is a UI to choose the files.}

\item{object}{An \linkS4class{ntsData} object to add the selected files.}

\item{copy}{Logical. Set to \code{TRUE} to copied the selected files
to the \code{path} of the \linkS4class{ntsData} object.
\code{FALSE} will add the original location of the files
to the \code{samples} table.
Note that if duplicate names exist, the process is aborted.}

\item{replicates}{Optional character vector with the same length as the number of MS files to add
with the identifier/name of the sample replicate group for each file.
The default is \code{NULL} which attempts  to group the replicate samples using
the characters before a last \emph{-} (hyphen) or \emph{_} (underscore) of the MS file names.
The replicate names can always be edited using the class method \link{sampleGroups<-}.
See \code{?"sampleGroups<-"} for more information.}

\item{polarity}{A vector specifying the ionization polarity of the files.
Possible values are \emph{positive} (the default) or \emph{negative} for positive or negative modes, respectively.
Note that it will overwrite the polarity of the \linkS4class{ntsData} object.}

\item{method}{A character vector with the method used for acquiring the MS data.
When different methods were used, the method should be given per file.
So the character vector must have the same length as the number of MS files to add.}
}
\value{
Returns the \linkS4class{ntsData} object including the added files
in the \code{samples} slot.
}
\description{
Adds mzML and/or mzXML files to an \linkS4class{ntsData} object.
}
\note{
When the \linkS4class{ntsData} has files, new files are added
below in the \code{samples}. If processed data is already present,
the new files will invalidate the structure.
Therefore, data should be processed again.
}
