% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_001_makeSetup.R
\name{addMetadata}
\alias{addMetadata}
\title{addMetadata}
\usage{
addMetadata(obj, var = NULL, varname = "var")
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object to integrate sample replicate group metadata.}

\item{var}{A \code{data.frame} or a \code{vector} with the same
row number or length as the number of sample replicate groups in the \linkS4class{ntsData}.}

\item{varname}{A character vector to name the column in the metadata
\code{data.frame} when metadata is a \code{vector}. The default is \emph{var}.}
}
\value{
An \linkS4class{ntsData} with metadata added to the slot metadata.
}
\description{
Adds sample replicate group metadata to an \linkS4class{ntsData} object.
}
