% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_001_makeSetup.R
\name{addMetadata}
\alias{addMetadata}
\title{addMetadata}
\usage{
addMetadata(sampleInfo, metadata)
}
\arguments{
\item{sampleInfo}{The sampleInfo object that is to be amended by metadata.}

\item{metadata}{A list or table containing the desired metadata to be added to the sampleInfo object.
Needs to have the same number of entries as their are samples in the sampleInfo object.}
}
\value{
Returns a dataframe with the added metadata columns.
}
\description{
Adds metadata and and additional information to the sample
Additional information could be concentration data or other known properties of the samples.
}
