% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_001_setupProject.R
\name{addMetadata}
\alias{addMetadata}
\title{addMetadata}
\usage{
addMetadata(object, var = NULL, varname = "var")
}
\arguments{
\item{object}{An \linkS4class{ntsData} object.}

\item{var}{A \code{data.frame}, \code{data.table} or a \code{vector} with the same
row number or length as the number of sample replicates in the \linkS4class{ntsData}.
For \code{data.frame} or \code{data.table}, multiple columns can be added.}

\item{varname}{A character vector to name the column in the metadata
\code{data.table} when metadata is a \code{vector}. The default is \emph{var}.}
}
\value{
An \linkS4class{ntsData} with metadata added to the slot metadata.
}
\description{
Adds metadata for each sample in an \linkS4class{ntsData} object.
If metadata already exists in the \linkS4class{ntsData} object,
the new variable/s are added as new column/s.
}
