% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_001_setupProject.R
\name{addRawInfo}
\alias{addRawInfo}
\title{addRawInfo}
\usage{
addRawInfo(object, samples = NULL)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object.}

\item{samples}{A character/numeric vector with the name/index of the
samples to get raw data information.}
}
\value{
An \linkS4class{ntsData} object with raw information added
per sample/file.
}
\description{
Adds raw info for each sample/file
to an \linkS4class{ntsData} object.
}
