% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_000_S4_parameterClasses.R
\name{AlteredCameraParam}
\alias{AlteredCameraParam}
\title{AlteredCameraParam}
\usage{
AlteredCameraParam(
  sigma = 5,
  perfwhm = 0.45,
  cor_eic_th = 0.85,
  calcCaS = TRUE,
  cor_exp_th = 0.85,
  pval = 0.05,
  validateIsotopePatterns = TRUE,
  ppmIsotopes = 50,
  mzabs = 0.01,
  noise = 350,
  searchAdducts = TRUE,
  ppmAdducts = 5,
  extendedList = TRUE
)
}
\arguments{
\item{sigma}{The multiplier of the standard deviation for grouping features
by retention time.}

\item{perfwhm}{Percentage of the overlapping FWHM to group features.}

\item{cor_eic_th}{Threshold for feature EIC correlation in each sample.}

\item{calcCaS}{Calculate correlation across samples.}

\item{cor_exp_th}{Threshold for intensity correlation across samples.}

\item{pval}{p-value threshold for testing correlation of significance.}

\item{validateIsotopePatterns}{Logical, set to \code{TRUE} for validating
the annotated isotopes with the \emph{kegg} database.}

\item{ppmIsotopes}{The expected mass deviation (in ppm) to find isotopes.}

\item{mzabs}{numeric.}

\item{noise}{numeric.}

\item{searchAdducts}{Logical, set to \code{TRUE} to screen for adducts after finding isotopes.}

\item{ppmAdducts}{The expected mass deviation (in ppm) to find adducts.}

\item{extendedList}{Logical, set to \code{TRUE} to use the extended list of
adducts. The default is \code{FALSE}.}
}
\value{
An \linkS4class{AlteredCameraParam} class object for annotation.
}
\description{
AlteredCameraParam
}
