% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_006_buildFeatureList.R
\name{buildFeatureList}
\alias{buildFeatureList}
\title{buildFeatureList}
\usage{
buildFeatureList(obj)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}
}
\value{

}
\description{
buildFeatureList
}
\examples{

}
