% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_004_makeFeatures.R
\name{buildFeatureList}
\alias{buildFeatureList}
\title{buildFeatureList}
\usage{
buildFeatureList(obj)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}
}
\value{
An \linkS4class{ntsData} object with updated features slot.
}
\description{
buildFeatureList
}
