% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_006_FeaturesMetadata.R
\name{calculateSNR}
\alias{calculateSNR}
\title{calculateSNR}
\usage{
calculateSNR(obj, ID = NULL, rtWindow = 120)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{ID}{Optionally, a character vector with the ID of the features to calculate the sn ratio.}

\item{rtWindow}{The time window (before and after), in seconds,
to find centroids for estimation of the noise level.}
}
\value{
An \linkS4class{ntsData} object with the noise, noise sd and sn ratio columns ammended to the
features slot.
}
\description{
Calculates the signal-to-noise (sn) ratio for features in a given \linkS4class{ntsData} object
or specified via the argument \code{ID}. The noise is extimated by the maximum intensity of the background
centroids, which belong to the same mass bin as the peaks in the feature but are taken from a predefined
(\code{rtWindow}) time window before and after the feature limits. The sn is then calculated by
dividind the maximum intensity of the feature in the \linkS4class{ntsData} object with the estimated noise.
Centroids that belong to other peaks within the same mass and time windows are excluded before noise estimation.
}
