% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_002_importRawData.R
\name{centroidProfileData}
\alias{centroidProfileData}
\title{centroidProfileData}
\usage{
centroidProfileData(
  rawData,
  smoothing = TRUE,
  refineMZ = TRUE,
  methodSmoothing = "SavitzkyGolay",
  halfWindowSize = 5,
  polynomialOrder = 4,
  methodRefineMz = "kNeighbors",
  k = 2,
  signalPercentage = 10,
  stopAtTwo = TRUE,
  save = TRUE,
  projPath = setup$projPath
)
}
\arguments{
\item{rawData}{The \linkS4class{OnDiskMSnExp} object with profile data for centroiding.}

\item{smoothing}{Logical, set to \code{TRUE} for applying smothing to the profile data.}

\item{refineMZ}{Logical, set to \code{TRUE} for applying \emph{m/z} refinement before centroiding.}

\item{methodSmoothing}{Method for smoothing.}

\item{halfWindowSize}{The window size of the filter.}

\item{polynomialOrder}{Polynomial order for smothing, default is 4.}

\item{methodRefineMz}{Method for refinement.}

\item{k}{Number of closest signals to the centroid.}

\item{signalPercentage}{Minimum signal percentage of centroids to refine \emph{m/z}.}

\item{stopAtTwo}{Logical, set to \code{TRUE} for allowed two consecutive equal or higher signals.
\code{FALSE} will stop when one equal or higher centroid is found.}

\item{save}{Logical, set to \code{TRUE} to save the object rawData in the \strong{rData} folder.
Note, that the profile files will be replaced by the produced centroided files.}

\item{projPath}{If save is \code{TRUE}, the saving location or project path should be given.
The default is the \code{projPath} in the object \code{setup}.}
}
\value{
Centroiding of profile data from mzML and mzXML files in an \linkS4class{OnDiskMSnExp} object.
}
\description{
Centroid profile MS data with additional smoothing and refinement of the m/z.
}
\examples{



}
