% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_008_checkInternalStandards.R
\name{checkIS}
\alias{checkIS}
\title{checkIS}
\usage{
checkIS(
  obj = NULL,
  targets = NULL,
  ppm = NULL,
  rtWindow = NULL,
  MS2param = NULL,
  recoveryFrom = NULL,
  exportResults = FALSE,
  save = FALSE
)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{targets}{The \linkS4class{suspectList} object with the information of
internal standards to be found.}

\item{ppm}{The mass deviation, in ppm, to screen for the internal standards.
The default is 10 ppm.}

\item{rtWindow}{The retention time deviation, in seconds,
to screen for the internal standards. The default is 30 seconds.}

\item{MS2param}{The parameters to extract MS2 information of each internal standard.}

\item{recoveryFrom}{A character vector with the name of the sample replicate group
to use for calculating the recovery.
The default is \code{NULL} to use the intControl in the \linkS4class{suspectList}.}

\item{exportResults}{Logical, set to \code{TRUE} for exporting the results
to the project folder.}

\item{save}{Logical, set to \code{TRUE} to save updated
\linkS4class{ntsData} object in the \strong{rdata} folder.
Note that \code{TRUE} overwrites the existing \linkS4class{ntsData} object.
Optionally, a character string can be given instead of \code{TRUE}
to be used as file name, avoiding overwriting.}
}
\value{
An \linkS4class{ntsData} object with results from evaluation of
internal standards in each sample replicate group added to the IS slot.
}
\description{
Evaluation of internal standards for each sample replicate group
within the \linkS4class{ntsData} object.
}
