% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_011_qualityControl.R
\name{checkQC}
\alias{checkQC}
\title{checkQC}
\usage{
checkQC(
  obj = NULL,
  targets = NULL,
  rtWindow = NULL,
  ppm = NULL,
  exportResults = FALSE,
  save = FALSE
)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object, containing QC samples.}

\item{targets}{The \linkS4class{suspectList} object with the target compounds
for quality control.}

\item{rtWindow}{The retention time deviation, in seconds,
to screen for the QC target standards. The default is 30 seconds.}

\item{ppm}{The mass deviation, in ppm, to screen for the QC target standards.
The default is 15 ppm.}

\item{exportResults}{Logical, set to \code{TRUE} for exporting the results to the project folder.}

\item{save}{Logical, set to \code{TRUE} to save updated
\linkS4class{ntsData} object in the \strong{rdata} folder.
Note that \code{TRUE} overwrites the existing \linkS4class{ntsData} object.
Optionally, a character string can be given instead of \code{TRUE}
to be used as file name, avoiding overwriting.}
}
\value{
The \linkS4class{ntsData} with quality control data and
evaluation results added to the slot \code{QC}.
}
\description{
Function to check QC samples in an \linkS4class{ntsData} object.
}
\details{
The \code{screeningList} template can be obtained as csv via \code{ntsIUTA::getScreeningListTemplate()}.
Add other details of the template.
}
