% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_009_qualityControl.R
\name{checkQC}
\alias{checkQC}
\title{checkQC}
\usage{
checkQC(
  rawQC = rawData,
  sampleInfo = setup$sampleInfo,
  screeningList = utils::choose.files(base::getwd(), "Select the QC screening list"),
  paramPeaks = NULL,
  paramPreGrouping = NULL,
  paramAlignment = NULL,
  paramGrouping = NULL,
  ppmForFillingGroups = 5,
  rtWindow = 30,
  ppmWindow = 15,
  polarity = "positive",
  plot = TRUE,
  save = FALSE,
  projPath = setup$projPath
)
}
\arguments{
\item{rawQC}{An \linkS4class{OnDiskMSnExp} object corresponding to the QC samples.
The function \code{filterFile} from \code{MSnbase} can be used to subset the \code{rawData}, such as \code{MSnbase::filterFile(rawData, file = 4:6)}.}

\item{sampleInfo}{The \code{sampleInfo} data frame generated by \code{\link{makeSetup}} function.}

\item{screeningList}{A \code{data.frame} with details from each standard in the QC samples.
See details for more information about the required data.frame structure.}

\item{paramPeaks}{The parameters for the choosen peak picking method.
See documentation of \code{\link[xcms]{chromatographic-peak-detection}} for more information.}

\item{paramPreGrouping}{If \code{\link[xcms]{adjustRtime}} is preformed with the method \code{PeakGroups},
a pre-grouping of peaks is required for alignment. The \code{paramPreGrouping} is the parameters obtained by the selected grouping method.
See documentation of \code{\link[xcms]{groupChromPeaks}} for more information.}

\item{paramAlignment}{The parameters for the choosen alignment method. See documentation of \code{\link[xcms]{adjustRtime}} for more information.}

\item{paramGrouping}{The parameters for the choosen grouping method.
See documentation of \code{\link[xcms]{groupChromPeaks}} for more information.}

\item{ppmForFillingGroups}{The mass (in ppm) to expand the \emph{mz} for filling missing peaks in imcomplete features.
See \code{\link[xcms]{fillChromPeaks}} for more information.}

\item{rtWindow}{The retention time deviation, in seconds, allowed to screen for QC reference standards.}

\item{ppmWindow}{The mass deviation, in ppm, allowed to screen for QC reference standards.}

\item{polarity}{The acquisition polarity of the QC samples. Possible values are \code{positive} or \code{negative}.}

\item{plot}{Logical, set to \code{TRUE} for plotting the results.}

\item{save}{Logical, set to \code{TRUE} for storing  the generated QC object in the rData folder.}

\item{projPath}{The \code{projPath} directory as defined in the \code{setup} object.}
}
\value{
A \code{list} containing the plots and a data.frame with the summary of the
retention time, mass and intensity deviations as well as the quality of the MS2 data.
}
\description{
Function to check QC samples in \code{rawData}.
}
\details{
The \code{screeningList} template can be obtained as csv via \code{ntsIUTA::getScreeningListTemplate()}.
Add other details of the template.
}
\examples{



}
