% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_002_wrapperFunctions.R
\name{createFeatures}
\alias{createFeatures}
\title{createFeatures}
\usage{
createFeatures(object, save = FALSE)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object.}

\item{save}{Logical, set to \code{TRUE} to save updated
\linkS4class{ntsData} object in the \strong{rdata} folder.
Note that \code{TRUE} overwrites the existing \linkS4class{ntsData} object.
Optionally, a character string can be given instead of \code{TRUE}
to be used as file name, avoiding overwriting.}
}
\value{
An \linkS4class{ntsData} object with peaks
grouped and annotated across samples as features.
}
\description{
Wrapper function for peak picking, grouping, filling, annotation
and creation of UFIs when settings are already defined in the \linkS4class{ntsParameters}.
}
\note{
The \linkS4class{ntsData} object should contain the parameter settings
required for creation and annotation of features.
}
