% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_002_importRawData.R
\name{extractEIC}
\alias{extractEIC}
\title{extractEIC}
\usage{
extractEIC(
  obj = NULL,
  fileIndex = NULL,
  mz = NULL,
  ppm = NULL,
  rt = NULL,
  rtWindow = NULL,
  rtUnit = "sec",
  msLevel = 1,
  normIntensity = FALSE
)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object with one or more files.}

\item{fileIndex}{The index of the file/s to extract the centroids or profile data.}

\item{mz}{Target \emph{m/z} to the EIC.}

\item{ppm}{The mass deviation to extract the data for the EIC in \code{ppm}.}

\item{rt}{The retention time in minutes or seconds, depending on the defined \code{rtUnit}, see below.
With Zero or \code{NULL} the complete retention time will be used.}

\item{rtWindow}{The time deviation to collect centroids or profile data. The time unit is the defined by \code{rtUnit}.
A time interval can be given with a length 2 vector, defining the minimum and maximum retention time.}

\item{rtUnit}{Possible entries are \code{min} or \code{sec}. The default is \code{sec}.}

\item{msLevel}{The MS level to extract the data. For the moment, only 1 is possible.}

\item{normIntensity}{Logical, set to \code{TRUE} for normalizing the intensity values between 1 and 0 for each file in the given \code{raw} object.}
}
\value{
A \code{data.frame} with the columns \code{file}, \code{rt}, \code{mz} and \code{i}
representing the mzML file index, the retention time, the \emph{m/z} and the intensity, respectively.
}
\description{
Extracts an ion chromatogram (EIC) from raw data of a specified \emph{m/z}.
}
\examples{

}
