% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{features}
\alias{features}
\title{features}
\usage{
features(object, ...)
}
\arguments{
\item{object}{An object to extract features.}

\item{...}{Other method dependent parameters.}
}
\value{
A data.frame with feature details.
}
\description{
features
}
