% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_011_findFragments.R
\name{findFragments}
\alias{findFragments}
\title{findFragments}
\usage{
findFragments(
  obj,
  targets = NULL,
  replicates = NULL,
  ID = NULL,
  title = NULL,
  ppm = 10,
  intMin = 10,
  MS2param = NULL
)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object with features for finding fragment in MS2 data.}

\item{targets}{A data frame with target values for matching fragments.
See details for more information.}

\item{replicates}{A character vector with the name of the sample replicate groups to use for finding fragments.}

\item{ID}{A character vector with the ID of the features of interest.}

\item{title}{Optional, title to use for the entry in the workflows slot of the \linkS4class{ntsData} object.}

\item{ppm}{The mass deviation, in ppm, for matching fragments.}

\item{intMin}{The minimum intensity of fragments.}

\item{MS2param}{A list of parameters to use for extracting MS2 data.}
}
\value{
An \linkS4class{ntsData} object with an \linkS4class{ntsFragmentsData} object
add to the workflows slot.
}
\description{
findFragments
}
