% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_001_supportTools.R
\name{getColors}
\alias{getColors}
\title{getColors}
\usage{
getColors(x, which = c("groups", "samples"))
}
\arguments{
\item{x}{An \linkS4class{OnDiskMSnExp} object with one or more files or the number of colors to be produced.}

\item{which}{Possible entries are \code{groups} and \code{samples} for getting group or samples colors, respectively.}
}
\value{
A vector of colors names according to the given \linkS4class{OnDiskMSnExp} object.
}
\description{
getColors
}
\examples{



}
