% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_001_supportTools.R
\name{getColors}
\alias{getColors}
\title{getColors}
\usage{
getColors(x, which = c("samples", "sampleGroups", "groups"))
}
\arguments{
\item{x}{An \linkS4class{ntsData} object with one or more files or the number of colours to be produced.}

\item{which}{Possible entries are \code{samples}, \code{sampleGroups} and \code{groups} for getting
individual sample, sample replicate group and individual group colours, respectively.}
}
\value{
A vector of colours, named according to a \linkS4class{ntsData} object if given.
}
\description{
getColors
}
