% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_002_importRawData.R
\name{importRawData}
\alias{importRawData}
\title{importRawData}
\usage{
importRawData(
  sampleInfo = sampleInfo,
  rtFilter = NULL,
  timeUnit = "min",
  msLevel = c(1, 2),
  centroidedData = TRUE,
  removeEmptySpectra = TRUE,
  save = FALSE,
  projPath = base::getwd()
)
}
\arguments{
\item{sampleInfo}{The sample \code{data.table} obtained by the \code{\link{setupProject}} function.
Note that the \code{sampleInfo} can be filtered beforehand to exclude unwanted mzML files or to redefine the name of replicate sample groups.}

\item{rtFilter}{A numeric vector with length 2 defining the minimum and maximum chromatographic retention time for the listed files respectively.
If the files have different time dimensions, \code{rtFilter} should be \code{NULL}.}

\item{timeUnit}{If \code{rtFilter} is given in minutes please keep the default value \code{min} otherwise change to \code{sec} for seconds.}

\item{msLevel}{The MS dimensions for the rtFilter to be applied. Default is for both MS1 and MS2 with \code{c(1,2)}.}

\item{centroidedData}{Logical, set to \code{TRUE} for mzML files with centroided data or \code{FALSE} for profile data.
\code{NA} will collect all the data from the mzML file.}

\item{removeEmptySpectra}{Logical, set to TRUE if empty spectra should be removed.
It is recommended to remove empty spectra as it may cause issues during creation of features.}

\item{save}{Logical, set to \code{TRUE} to save the object rawData in the \strong{rData} folder.
If \strong{rData} folder does not exist in given \code{projPath} from \code{setup}, the folder will be created.}

\item{projPath}{If save is \code{TRUE}, the saving location or project path should be given.
The default is the \code{projPath} in the object \code{setup}.}
}
\value{
The output is a standard \linkS4class{OnDiskMSnExp} object,
which can be used within the workflow of \pkg{ntsIUTA} but also within \pkg{Bioconductor} packages, such as \pkg{xcms}.
}
\description{
Function to import MS data from the mzML files listed in \code{sampleInfo} from the \code{setupProject} object.
Files should contain centroided spectra.
The function uses the function \code{\link[MSnbase]{readMSData}} from the \code{MSnbase} package to read the mzML files.
}
\examples{
sampleInfo <- ntsIUTA::setupProject(projPath = system.file(package = "ntsIUTA", dir = "extdata"), save = FALSE)
rawData <- ntsIUTA::importRawData(sampleInfo[1,], save = FALSE, centroidedData = TRUE)
rawData

}
\references{
\insertRef{MSnbase1}{ntsIUTA}
\insertRef{MSnbase2}{ntsIUTA}
}
