% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_004_makeFeatures.R
\name{makeFeatures}
\alias{makeFeatures}
\title{makeFeatures}
\usage{
makeFeatures(
  obj = NULL,
  algorithm = NULL,
  paramGrouping = NULL,
  recursive = TRUE,
  paramFill = NULL,
  save = TRUE
)
}
\arguments{
\item{obj}{A \linkS4class{ntsData} object containing peaks
for one or more samples.}

\item{algorithm}{The algorithm to use for peak alignment and grouping.
One of "xcms3" (the default) or "openms".
When "openms" the \pkg{patRoon} package is used.}

\item{paramGrouping}{The parameters for the chosen grouping method.
See documentation of \code{\link[xcms]{groupChromPeaks}} or
\code{\link[patRoon]{groupFeatures}} for more information.}

\item{recursive}{Logical, set to \code{TRUE} for applying recursive
integration for samples with missing peaks.}

\item{paramFill}{An object of class \code{FillChromPeaksParam}
or \code{ChromPeakAreaParam} containing the parameters
to apply the recursive integration.
#' See \code{?\link[xcms]{fillChromPeaks}} for more information.}

\item{save}{Logical, set to \code{TRUE} to save updated
\linkS4class{ntsData} object in the \strong{rdata} folder.
Note that \code{TRUE} overwrites the existing \linkS4class{ntsData} object.
Optionally, a character string can be given instead of \code{TRUE}
to be used as file name, avoiding overwriting.}
}
\value{
An \linkS4class{ntsData} object
containing peaks aligned and grouped as features across samples accessible
in the slot \code{features}.

 @references
\insertRef{Helmus2021}{ntsIUTA}
\insertRef{xcms1}{ntsIUTA}
\insertRef{xcms2}{ntsIUTA}
\insertRef{xcms3}{ntsIUTA}
}
\description{
The \code{makeFeatures} consists of alignment and grouping of
chromatographic peaks across samples. The function uses algorithms
from the package \pkg{xcms} or from \pkg{patRoon} based on "openms".
Additionally, a recursive integration for samples with missing peaks
can be applied using the function \code{\link[xcms]{fillChromPeaks}}
from the \pkg{xcms} package.
}
