% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_000_S4_ntsdata.R
\docType{class}
\name{ntsData-class}
\alias{ntsData-class}
\alias{show,ntsData-method}
\alias{samples,ntsData-method}
\alias{sampleGroups,ntsData-method}
\alias{sampleGroups<-,ntsData-method}
\alias{blanks,ntsData-method}
\alias{blanks<-,ntsData-method}
\alias{[,ntsData,ANY,missing,missing-method}
\alias{QC,ntsData-method}
\alias{QC<-,ntsData-method}
\title{ntsData}
\usage{
\S4method{show}{ntsData}(object)

\S4method{samples}{ntsData}(object)

\S4method{sampleGroups}{ntsData}(object)

\S4method{sampleGroups}{ntsData}(object) <- value

\S4method{blanks}{ntsData}(object)

\S4method{blanks}{ntsData}(object) <- value

\S4method{[}{ntsData,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{QC}{ntsData}(object)

\S4method{QC}{ntsData}(object) <- value
}
\arguments{
\item{\dots}{Ignored.}
}
\value{
An S4 class object named \linkS4class{ntsData}.
}
\description{
S4 class object to organize and store processed data
within the \pkg{ntsIUTA} package.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Informative printing.

\item \code{samples}: Getter for samples.

\item \code{sampleGroups}: Getter for sample replicate groups.

\item \code{sampleGroups<-}: Setter for sample replicate groups.

\item \code{blanks}: Getter for blank replicate groups.

\item \code{blanks<-}: Setter for blank replicate groups.

\item \code{[}: Subset on samples.

\item \code{QC}: Getter for QC replicate sample groups.

\item \code{QC<-}: Setter for QC replicate sample groups.
}}

\section{Slots}{

\describe{
\item{\code{title}}{A character string with the project title.}

\item{\code{info}}{A list to store project details.
The first position is for description of the project.}

\item{\code{path}}{A character string with the project path.
as defined with \code{\link{setupProject}}.}

\item{\code{date}}{The project date.}

\item{\code{polarity}}{A string character defining the polarity mode of
the MS files in the \linkS4class{ntsData} object.
Possible values are "positive" or "negative".}

\item{\code{samples}}{A \code{data.frame} as obtained by \code{\link{setupProject}}.}

\item{\code{metadata}}{A \code{data.frame} with the same number of rows
as the \code{samples} containing metadata to support data interpretation.}

\item{\code{parameters}}{A list containing process parameters for the workflow steps.}

\item{\code{QC}}{Slot for assigning the QC sample replicate group and results.
Note that the assigned group is not included in the basic workflow
but only used for quality control using the function \code{\link{qualityControl}}.}

\item{\code{MSnExp}}{An \linkS4class{OnDiskMSnExp} as generated by \code{\link{setupProject}}.}

\item{\code{patdata}}{A \linkS4class{features} or \linkS4class{featureGroups}
object derived from the basic NTS workflow
(peak picking, alignment and grouping).}

\item{\code{peaks}}{A list with a data.frame containing the peak information
for each sample in the slot \code{samples}.
same length as number of rows with peak information.}

\item{\code{features}}{data.frame.}
}}

\note{
Slot \code{MSnExp} is used to save the \linkS4class{OnDiskMSnExp}
object obtained by \pkg{MSnbase} during the \code{\link{setupProject}}.
Slot \code{patdata} will contain the object \linkS4class{features} or
\linkS4class{featureGroups} generated from the \pkg{patRoon} package.
Both objects can be used within native functions of \pkg{MSnbase} and
\pkg{patRoon}.
}
\examples{

}
