% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_000_S4_ntsdata.R
\docType{class}
\name{ntsData-class}
\alias{ntsData-class}
\alias{show,ntsData-method}
\alias{pullSamples,ntsData-method}
\alias{pullSamplegroups,ntsData-method}
\alias{pushSamplegroups,ntsData-method}
\alias{pushSamplegroups<-,ntsData-method}
\alias{pullBlanks,ntsData-method}
\alias{pushBlanks,ntsData-method}
\alias{pushBlanks<-,ntsData-method}
\alias{pullQC,ntsData-method}
\alias{pushQC<-,ntsData-method}
\alias{[,ntsData,ANY,missing,missing-method}
\alias{pullPeaks,ntsData-method}
\alias{pullFeatures,ntsData-method}
\title{ntsData}
\usage{
\S4method{show}{ntsData}(object)

\S4method{pullSamples}{ntsData}(object)

\S4method{pullSamplegroups}{ntsData}(object)

\S4method{pushSamplegroups}{ntsData}(object, value)

\S4method{pushSamplegroups}{ntsData}(object) <- value

\S4method{pullBlanks}{ntsData}(object)

\S4method{pushBlanks}{ntsData}(object, value)

\S4method{pushBlanks}{ntsData}(object) <- value

\S4method{pullQC}{ntsData}(object)

\S4method{pushQC}{ntsData}(object) <- value

\S4method{[}{ntsData,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{pullPeaks}{ntsData}(
  obj,
  fileIndex = NULL,
  ID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  rtWindow = NULL,
  rtUnit = "sec"
)

\S4method{pullFeatures}{ntsData}(
  obj,
  fileIndex = NULL,
  ID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  rtWindow = NULL,
  rtUnit = "sec"
)
}
\arguments{
\item{\dots}{Ignored.}
}
\value{
An S4 class object named \linkS4class{ntsData}.
}
\description{
S4 class object to organize and store processed data
within the \pkg{ntsIUTA} package.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Informative printing.

\item \code{pullSamples}: Getter for samples.

\item \code{pullSamplegroups}: Getter for sample replicate groups.

\item \code{pushSamplegroups}: Setter for sample replicate groups.

\item \code{pushSamplegroups<-}: Setter for sample replicate groups.

\item \code{pullBlanks}: Getter for blank replicate groups.

\item \code{pushBlanks}: Setter for blank replicate groups.

\item \code{pushBlanks<-}: Setter for blank replicate groups.

\item \code{pullQC}: Getter for QC replicate sample groups.

\item \code{pushQC<-}: Setter for QC replicate sample groups.

\item \code{[}: Subset on samples.

\item \code{pullPeaks}: Getter for chromatographic peaks.

\item \code{pullFeatures}: Getter for features (i.e., grouped peaks).
}}

\section{Slots}{

\describe{
\item{\code{title}}{A character string with the project title.}

\item{\code{info}}{A list to store project details.
The first position is for description of the project.}

\item{\code{path}}{A character string with the project path.
as defined with \code{\link{setupProject}}.}

\item{\code{date}}{The project date.}

\item{\code{polarity}}{A string character defining the polarity mode of
the MS files in the \linkS4class{ntsData} object.
Possible values are "positive" or "negative".}

\item{\code{samples}}{A \code{data.frame} as obtained by \code{\link{setupProject}}.}

\item{\code{metadata}}{A \code{data.frame} with the same number of rows
as the \code{samples} containing metadata to support data interpretation.}

\item{\code{algorithms}}{A list of algorithms used for the basic workflow steps,
which are \emph{Peak Picking}, \emph{Make Features} and \emph{Annotation}.}

\item{\code{parameters}}{A list containing process parameters for the workflow steps.}

\item{\code{QC}}{Slot for assigning the QC sample replicate group and results.
Note that the assigned group is not included in the basic workflow
but only used for quality control, using the function \code{\link{checkQC}},
and for optimization of workflow parameters. A screening list with target substances
must be provided for running quality control.}

\item{\code{MSnExp}}{An \linkS4class{OnDiskMSnExp} as generated by \code{\link{setupProject}}.}

\item{\code{patdata}}{A \linkS4class{features} or \linkS4class{featureGroups}
object derived from the basic NTS workflow
(peak picking, alignment and grouping).}

\item{\code{peaks}}{A list with a data.frame containing the peak information
for each sample in the slot \code{samples}.
same length as number of rows with peak information.}

\item{\code{features}}{data.frame.}

\item{\code{annotation}}{Slot for storage of annotation of isotopes and adducts and
source fragments. Features with similar chromatographic behaviour
are grouped by components. Isotopologues are grouped by \code{isonumber} and
are created for each sample replicate group. Similarly, adducts are searched
for each sample replicate group and share the same molecular ion (\code{Mion}).
The functionality of the package \pkg{CAMERA} was adapted to work better for
environmental analyses, where sample replicate groups are, in principle,
very different from each other. When specified other functionalities
from the package \pkg{patRoon} can be used but the default
is the package \pkg{CAMERA} directly. The slot is composed of a list of length 2,
where the first entry is a data frame with summarized results for each feature
and the second is the raw data from the \pkg{CAMERA} package.}
}}

\note{
Slot \code{MSnExp} is used to save the \linkS4class{OnDiskMSnExp}
object obtained by \pkg{MSnbase} during the \code{\link{setupProject}}.
Slot \code{patdata} will contain the object \linkS4class{features} or
\linkS4class{featureGroups} generated from the \pkg{patRoon} package.
Both objects can be used within native functions of \pkg{MSnbase} and
\pkg{patRoon}.
}
\examples{

}
