% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R, R/S4_methods_001_ntsdata.R,
%   R/S4_methods_002_parameters.R
\docType{class}
\name{ntsData-class}
\alias{ntsData-class}
\alias{projectInfo,ntsData-method}
\alias{path,ntsData-method}
\alias{samples,ntsData-method}
\alias{replicates,ntsData-method}
\alias{replicates<-,ntsData-method}
\alias{blanks,ntsData-method}
\alias{blanks<-,ntsData-method}
\alias{acquisitionMethods,ntsData-method}
\alias{acquisitionMethods<-,ntsData-method}
\alias{polarity,ntsData-method}
\alias{polarity<-,ntsData-method}
\alias{metadata,ntsData-method}
\alias{QC,ntsData-method}
\alias{QC<-,ntsData-method}
\alias{[,ntsData,ANY,missing,missing-method}
\alias{[,ntsData,ANY,ANY,missing-method}
\alias{peaks,ntsData-method}
\alias{features,ntsData-method}
\alias{components,ntsData-method}
\alias{show,ntsData-method}
\alias{peakPickingParameters,ntsData,missing,missing-method}
\alias{peakPickingParameters,ntsData,character,ANY-method}
\alias{peakGroupingParameters,ntsData,missing,missing-method}
\alias{peakGroupingParameters,ntsData,character,ANY-method}
\alias{fillMissingParameters,ntsData,missing,missing-method}
\alias{fillMissingParameters,ntsData,character,ANY-method}
\alias{annotationParameters,ntsData,missing,missing-method}
\alias{annotationParameters,ntsData,character,ANY-method}
\title{ntsData}
\usage{
\S4method{projectInfo}{ntsData}(object, title = NULL, description = NULL, date = NULL)

\S4method{path}{ntsData}(object)

\S4method{samples}{ntsData}(object)

\S4method{replicates}{ntsData}(object)

\S4method{replicates}{ntsData}(object) <- value

\S4method{blanks}{ntsData}(object)

\S4method{blanks}{ntsData}(object) <- value

\S4method{acquisitionMethods}{ntsData}(object)

\S4method{acquisitionMethods}{ntsData}(object) <- value

\S4method{polarity}{ntsData}(object, groupby)

\S4method{polarity}{ntsData}(object) <- value

\S4method{metadata}{ntsData}(x, varname)

\S4method{QC}{ntsData}(object)

\S4method{QC}{ntsData}(object, remove = FALSE, nametype = "replicates") <- value

\S4method{[}{ntsData,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{ntsData,ANY,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{peaks}{ntsData}(
  object,
  samples = NULL,
  ID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  rtWindow = NULL,
  rtUnit = "sec"
)

\S4method{features}{ntsData}(
  object,
  samples = NULL,
  ID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  rtWindow = NULL,
  rtUnit = "sec"
)

\S4method{components}{ntsData}(
  object,
  samples = NULL,
  ID = NULL,
  mz = NULL,
  ppm = 5,
  rt = NULL,
  rtWindow = 1,
  rtUnit = "sec",
  compNumber = NULL,
  entireComponents = TRUE,
  onlyAnnotated = FALSE,
  onlyRelated = TRUE
)

\S4method{show}{ntsData}(object)

\S4method{peakPickingParameters}{ntsData,missing,missing}(object)

\S4method{peakPickingParameters}{ntsData,character,ANY}(object, algorithm, param)

\S4method{peakGroupingParameters}{ntsData,missing,missing}(object)

\S4method{peakGroupingParameters}{ntsData,character,ANY}(object, algorithm, param)

\S4method{fillMissingParameters}{ntsData,missing,missing}(object)

\S4method{fillMissingParameters}{ntsData,character,ANY}(object, algorithm, param)

\S4method{annotationParameters}{ntsData,missing,missing}(object)

\S4method{annotationParameters}{ntsData,character,ANY}(object, algorithm, param)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object.}

\item{title}{The project title.}

\item{description}{The project description.}

\item{date}{The project date.}

\item{value}{A character vector with the names of the samples or sample replicate name/s
to be used for QC.}

\item{groupby}{Either \code{samples} (the default) or \code{replicates} as character string
to return the polarites either for each sample or each replicate.}

\item{x}{An \linkS4class{ntsData} object.}

\item{varname}{The name of the variables to extract.}

\item{remove}{A logical value. When \code{TRUE}, specified sample or replicate names
are moved from the QC slot to the samples slot of the \linkS4class{ntsData} object.}

\item{nametype}{A character string specifying if sample or replicate names should be used.
Posiible values are \emph{samples} or \emph{replicates} (the default).}

\item{i}{The indice/s or name/s of the samples to keep in the \code{x} object.}

\item{j}{Ignored.}

\item{\dots}{Ignored.}

\item{drop}{Ignored.}

\item{samples}{The indice/s or name/s of samples to keep in the \code{object}.}

\item{ID}{The ID of features of interest.}

\item{mz}{Alternatively to \code{ID}, the \emph{m/z} of interest.
can be of length two, defining a mass range.}

\item{ppm}{The mass deviation, in ppm, of a given \code{mz}.}

\item{rt}{The retention time to find features.}

\item{rtWindow}{The time deviation. Can be of length two, defining a time range.}

\item{rtUnit}{The unit of the time arguments. Possible values are "sec" and "min".}

\item{compNumber}{Alternatively, the component number to find features.}

\item{entireComponents}{Logical, set to \code{TRUE} to extract all features
from the represented components.}

\item{onlyAnnotated}{Logical, set to \code{TRUE} to extract only annotated features.}

\item{onlyRelated}{Logical, set to \code{TRUE} to extract only features that are related
to the features of interest.}

\item{algorithm}{A character vector with the name of the algorithm to be applied.}

\item{param}{A list with parameters matching the defined algorithm.}
}
\value{
An S4 class object named \linkS4class{ntsData}.

An \linkS4class{ntsData} object with updated project information.

A character vector with project path.

A character vector with sample names.

A character vector with sample replicate names.

Adds the value vector with sample replicate names
to the replicate column of the samples data table in the object.

A character vector with blank sample replicate names.

Adds the value vector with blank sample replicate names
to the blank column of the samples data table in the object.

A vector with acquisition method names used for each sample
or when only one method is present, the single acquisition method name.

Adds the \code{value} vector with acquisition method names
to the method column of the samples data table in the object.

A vector with the polarity mode of each sample (i.e., file).

Adds the \code{value} vector with polarity modes
to the polarity column of the samples data table in the object.

A data table with the requested metadata.
}
\description{
A S4 class object to store project processed data within the \pkg{ntsIUTA} package.
}
\section{Methods (by generic)}{
\itemize{
\item \code{projectInfo}: setter and getter for project information.

\item \code{path}: Getter for project path.

\item \code{samples}: Getter for sample names.

\item \code{replicates}: Getter for sample replicate names.

\item \code{replicates<-}: Setter for sample replicate names.

\item \code{blanks}: Getter for blank replicate names.

\item \code{blanks<-}: Setter for blank replicate groups.

\item \code{acquisitionMethods}: Getter for acquisition method names.

\item \code{acquisitionMethods<-}: Setter for acquisition method names.

\item \code{polarity}: Getter for the polarity of each sample/replicate.

\item \code{polarity<-}: Setter for the polarity mode of the samples (i.e., files).

\item \code{metadata}: Getter for metadata.

\item \code{QC}: Getter for QC samples data table.

\item \code{QC<-}: Setter for QC samples or sample replicates.

\item \code{[}: Subset on samples, using sample index or name.

\item \code{[}: Subset on samples, using sample index or name.

\item \code{peaks}: Getter for chromatographic peaks.

\item \code{features}: Getter for features (i.e., grouped peaks).

\item \code{components}: Getter for components (i.e., annotated features).

\item \code{show}: Informative printing of the \linkS4class{ntsData} object.

\item \code{peakPickingParameters}: Getter for peak picking parameters.

\item \code{peakPickingParameters}: Setter for peak picking parameters.

\item \code{peakGroupingParameters}: Getter for peak grouping parameters.

\item \code{peakGroupingParameters}: Setter for peak grouping parameters.

\item \code{fillMissingParameters}: Getter for fill missing parameters.

\item \code{fillMissingParameters}: Setter for fill missing parameters.

\item \code{annotationParameters}: Getter for annotation parameters.

\item \code{annotationParameters}: Setter for annotation parameters.
}}

\section{Slots}{

\describe{
\item{\code{title}}{A character string with the project title.}

\item{\code{description}}{A character string with the project description.}

\item{\code{date}}{The project date.}

\item{\code{path}}{A character string with the project path.}

\item{\code{samples}}{A data.table with six columns:
file (character string with the file path),
sample (character string with the file name),
replicate (character string with the assigned sample replicate group),
blank (character string with associated blank replicate),
polarity (the polarity mode of the respective file; possible values are "positive" or "negative") and
method (a character string with the method used to acquire the data file).}

\item{\code{metadata}}{A data.table with the same number of rows
as the unique replicates in \code{samples}, containing metadata for each replicate added as extra columns.}

\item{\code{parameters}}{A \linkS4class{paramList} object containing process parameters for the basic workflow.
See \code{?"paranList-class"} for more information.}

\item{\code{QC}}{A \linkS4class{qcData} object used for quality control. More information in \code{?"qcData-class"}.}

\item{\code{pat}}{A list of \linkS4class{features} or \linkS4class{featureGroups} objects
for each sample replicate derived from the basic workflow
(peak picking, alignment and grouping) using the \pkg{patRoon} package.}

\item{\code{peaks}}{A list of data.tables with peaks for each sample.}

\item{\code{features}}{A list of data.tables with the features for each replicate.}

\item{\code{unified}}{A data.table with unified features across different replicates.}

\item{\code{filters}}{A list of applied filters to the unified features data.table.}

\item{\code{removed}}{A data table with the removed unified features.}

\item{\code{workflows}}{A list of objects inherent of downstream data processing steps, such as
suspect screening.}
}}

\note{
The slot \code{pat} contains the object \linkS4class{features} or
\linkS4class{featureGroups} which can be used within native functions of the \pkg{patRoon} package.

When no arguments are given in the call for \code{projectInfo} a list
with the title, description and date is returned instead of the \code{object}.
}
