% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R, R/S4_methods_001_ntsdata.R,
%   R/S4_methods_002_parameters.R
\docType{class}
\name{ntsData-class}
\alias{ntsData-class}
\alias{show,ntsData-method}
\alias{samples,ntsData-method}
\alias{sampleGroups,ntsData-method}
\alias{sampleGroups<-,ntsData-method}
\alias{blanks,ntsData-method}
\alias{blanks<-,ntsData-method}
\alias{QC,ntsData-method}
\alias{QC<-,ntsData-method}
\alias{[,ntsData,ANY,missing,missing-method}
\alias{[,ntsData,ANY,ANY,missing-method}
\alias{peaks,ntsData-method}
\alias{features,ntsData-method}
\alias{components,ntsData-method}
\alias{peakPickingParameters,ntsData,missing,missing-method}
\alias{peakPickingParameters,ntsData,character,ANY-method}
\alias{peakGroupingParameters,ntsData,missing,missing-method}
\alias{peakGroupingParameters,ntsData,character,ANY-method}
\alias{fillMissingParameters,ntsData,missing,missing-method}
\alias{fillMissingParameters,ntsData,character,ANY-method}
\alias{annotationParameters,ntsData,missing,missing-method}
\alias{annotationParameters,ntsData,character,ANY-method}
\title{ntsData}
\usage{
\S4method{show}{ntsData}(object)

\S4method{samples}{ntsData}(object)

\S4method{sampleGroups}{ntsData}(object)

\S4method{sampleGroups}{ntsData}(object) <- value

\S4method{blanks}{ntsData}(object)

\S4method{blanks}{ntsData}(object) <- value

\S4method{QC}{ntsData}(object)

\S4method{QC}{ntsData}(object) <- value

\S4method{[}{ntsData,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{ntsData,ANY,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{peaks}{ntsData}(
  object,
  samples = NULL,
  ID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  rtWindow = NULL,
  rtUnit = "sec"
)

\S4method{features}{ntsData}(
  object,
  samples = NULL,
  ID = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  rtWindow = NULL,
  rtUnit = "sec"
)

\S4method{components}{ntsData}(
  object,
  samples = NULL,
  ID = NULL,
  mz = NULL,
  ppm = 5,
  rt = NULL,
  rtWindow = 1,
  rtUnit = "sec",
  compNumber = NULL,
  entireComponents = TRUE,
  onlyAnnotated = FALSE,
  onlyRelated = TRUE
)

\S4method{peakPickingParameters}{ntsData,missing,missing}(object)

\S4method{peakPickingParameters}{ntsData,character,ANY}(object, algorithm, param)

\S4method{peakGroupingParameters}{ntsData,missing,missing}(object)

\S4method{peakGroupingParameters}{ntsData,character,ANY}(object, algorithm, param)

\S4method{fillMissingParameters}{ntsData,missing,missing}(object)

\S4method{fillMissingParameters}{ntsData,character,ANY}(object, algorithm, param)

\S4method{annotationParameters}{ntsData,missing,missing}(object)

\S4method{annotationParameters}{ntsData,character,ANY}(object, algorithm, param)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object.}

\item{value}{A character vector with the names of the sample replicate group/s
to be used for QC.}

\item{x}{An \linkS4class{ntsData} object.}

\item{i}{The indice/s or name/s of the samples to keep in the \code{x} object.}

\item{j}{Ignored.}

\item{\dots}{Ignored.}

\item{drop}{Ignored.}

\item{samples}{The indice/s or name/s of samples to keep in the \code{object}.}

\item{ID}{The ID of features of interest.}

\item{mz}{Alternatively to \code{ID}, the \emph{m/z} of interest.
can be of length two, defining a mass range.}

\item{ppm}{The mass deviation, in ppm, of a given \code{mz}.}

\item{rt}{The retention time to find features.}

\item{rtWindow}{The time deviation. Can be of length two, defining a time range.}

\item{rtUnit}{The unit of the time arguments. Possible values are "sec" and "min".}

\item{compNumber}{Alternatively, the component number to find features.}

\item{entireComponents}{Logical, set to \code{TRUE} to extract all features
from the represented components.}

\item{onlyAnnotated}{Logical, set to \code{TRUE} to extract only annotated features.}

\item{onlyRelated}{Logical, set to \code{TRUE} to extract only features that are related
to the features of interest.}

\item{algorithm}{A character vector with the name of the algorithm to be applied.}

\item{param}{A list with parameters matching the defined algorithm.}
}
\value{
An S4 class object named \linkS4class{ntsData}.
}
\description{
S4 class object to organize and store project processed data
within the \pkg{ntsIUTA} package.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Informative printing.

\item \code{samples}: Getter for samples.

\item \code{sampleGroups}: Getter for sample replicate groups.

\item \code{sampleGroups<-}: Setter for sample replicate groups.

\item \code{blanks}: Getter for blank replicate groups.

\item \code{blanks<-}: Setter for blank replicate groups.

\item \code{QC}: Getter for QC replicate sample groups.

\item \code{QC<-}: Setter for QC replicate sample groups.

\item \code{[}: Subset on samples, using sample index or name.

\item \code{[}: Subset on samples, using sample index or name.

\item \code{peaks}: Getter for chromatographic peaks.

\item \code{features}: Getter for features (i.e., grouped peaks).

\item \code{components}: Getter for components (i.e., annotated features).

\item \code{peakPickingParameters}: Getter for peak picking parameters.

\item \code{peakPickingParameters}: Setter for peak picking parameters.

\item \code{peakGroupingParameters}: Getter for peak grouping parameters.

\item \code{peakGroupingParameters}: Setter for peak grouping parameters.

\item \code{fillMissingParameters}: Getter for fill missing parameters.

\item \code{fillMissingParameters}: Setter for fill missing parameters.

\item \code{annotationParameters}: Getter for annotation parameters.

\item \code{annotationParameters}: Setter for annotation parameters.
}}

\section{Slots}{

\describe{
\item{\code{title}}{A character string with the project title.}

\item{\code{info}}{A list to store project details.
The first position is for description of the project.}

\item{\code{path}}{A character string with the project path.
as defined with \code{\link{setupProject}}.}

\item{\code{date}}{The project date.}

\item{\code{polarity}}{A string character defining the polarity mode of
the MS files in the \linkS4class{ntsData} object.
Possible values are "positive" or "negative".}

\item{\code{samples}}{A \code{data.frame} as obtained by \code{\link{setupProject}}.}

\item{\code{metadata}}{A \code{data.frame} with the same number of rows
as the \code{samples} containing metadata to support data interpretation.}

\item{\code{parameters}}{A list containing process parameters for the workflow steps.}

\item{\code{QC}}{Slot for assigning the QC sample replicate group and results.
Note that the assigned group is not included in the basic workflow
but only used for quality control, using the function \code{\link{checkQC}},
and for optimization of workflow parameters. A screening list with target substances
must be provided for running quality control.}

\item{\code{MSnExp}}{An \linkS4class{OnDiskMSnExp} as generated by \code{\link{setupProject}}.}

\item{\code{patdata}}{A \linkS4class{features} or \linkS4class{featureGroups}
object derived from the basic NTS workflow
(peak picking, alignment and grouping).}

\item{\code{peaks}}{A list with a data.frame containing the peak information
for each sample in the slot \code{samples}.}

\item{\code{features}}{A data.frame with the features information,
including feature quality data and filtering information.}

\item{\code{annotation}}{Slot for storage of annotation of isotopes and adducts and
source fragments. Features with similar chromatographic behaviour
are grouped by components. Isotopologues are grouped by \code{isonumber} and
are created for each sample replicate group. Similarly, adducts are searched
for each sample replicate group and share the same molecular ion (\code{Mion}).
The functionality of the package \pkg{CAMERA} was adapted to work better for
environmental analyses, where sample replicate groups are, in principle,
very different from each other. When specified other functionalities
from the package \pkg{patRoon} can be used but the default
is the package \pkg{CAMERA} directly. The slot is composed of a list of length 2,
where the first entry is a data frame with summarized results for each feature
and the second is the raw data from the \pkg{CAMERA} package.}

\item{\code{IS}}{Slot to store the list of internal standards (IS) and the results
from IS control of the samples in the \code{ntsData} object.}

\item{\code{filters}}{A list of applied filters to the features data.frame.}

\item{\code{removed}}{A data frame with the removed features.}

\item{\code{workflows}}{A list of objects inherent of downstream workflows, such as
suspect screening.}
}}

\note{
Slot \code{MSnExp} is used to save the \linkS4class{OnDiskMSnExp}
object obtained by \pkg{MSnbase} during the \code{\link{setupProject}}.
Slot \code{patdata} will contain the object \linkS4class{features} or
\linkS4class{featureGroups} generated from the \pkg{patRoon} package.
Both objects can be used within native functions of \pkg{MSnbase} and
\pkg{patRoon}.
}
