% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{paramList-class}
\alias{paramList-class}
\title{paramList}
\value{
A \linkS4class{paramList} object to be used for
the basic workflow of \pkg{ntsIUTA}.
}
\description{
paramList
}
\section{Slots}{

\describe{
\item{\code{peakPicking}}{A paramSet object for performing peak picking.}

\item{\code{peakGrouping}}{A paramSet object for grouping peaks across samples.}

\item{\code{fillMissing}}{A paramSet object for performing recursive peak integration.}

\item{\code{annotation}}{A paramSet object for annotation of peaks and features in the \linkS4class{ntsData} object}

\item{\code{MS2}}{An \linkS4class{MS2param} object with settings for extraction of MS2 data.}
}}

