% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_005_peakFilling.R
\name{peakFilling}
\alias{peakFilling}
\title{peakFilling}
\usage{
peakFilling(object, algorithm = NULL, settings = NULL, save = TRUE)
}
\arguments{
\item{object}{An \linkS4class{ntsData} object containing features.}

\item{algorithm}{A character vector defining the algorithm to be used for filling peaks.}

\item{settings}{A list with parameters according to the defined algorithm.
When the algorithm is set to \emph{xcms3}, the settings are the S4 class objects
\linkS4class{FillChromPeaksParam} or \linkS4class{ChromPeakAreaParam}.
See \code{?\link[xcms]{fillChromPeaks}} for more information.}
}
\value{
An \linkS4class{XCMSnExp} object including filled missing peaks.
}
\description{
Recursive integration for filling missing peaks within each feature.
The function \code{\link[xcms]{fillChromPeaks}} from \pkg{xcms} package can be used,
giving the respectives parameters with the argument \code{settings}.
}
\references{
\insertRef{xcms1}{ntsIUTA}
\insertRef{xcms2}{ntsIUTA}
\insertRef{xcms3}{ntsIUTA}
}
