% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{peakGroupingParameters}
\alias{peakGroupingParameters}
\title{peakGroupingParameters}
\usage{
peakGroupingParameters(object, algorithm, param)
}
\arguments{
\item{object}{An object to add or get peak grouping parameters.}

\item{algorithm}{Algorithm for the set of parameters.}

\item{param}{List of parameters.}
}
\value{
A list of parameters or the object with added parameters.
}
\description{
peakGroupingParameters
}
