% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_003_peakPicking.R
\name{peakPicking}
\alias{peakPicking}
\title{peakPicking}
\usage{
peakPicking(obj = NULL, algorithm = "xcms3", param = NULL, save = TRUE)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{algorithm}{The algorithm to use for peak picking.
One of "xcms3" (the default), "xcms", "openms" or "envipick".}

\item{param}{List of arguments for the specified algorithm.
See \code{\link[patRoon]{findFeatures}} for more information.
For instance, for "xcms3" as \code{algorithm}, \code{param} should be given
with the method parameters for the peak finding.
See \href{https://rdrr.io/bioc/xcms/man/chromatographic-peak-detection.html}{\code{xcms::findChromPeaks}}
for more information.}

\item{save}{Logical, set to \code{TRUE} to save updated
\linkS4class{ntsData} object in the \strong{rdata} folder.
Note that \code{TRUE} overwrites the existing \linkS4class{ntsData} object.
Optionally, a character string can be given instead of \code{TRUE}
to be used as file name, avoiding overwriting.}
}
\value{
An \linkS4class{ntsData} object containing a \linkS4class{features}
object in the slot \code{patdata}.
}
\description{
Finds chromatographic peaks from centroided MS data
in an \linkS4class{ntsData} object.
The \code{peakPicking} function can use both \pkg{xcms}
and \pkg{patRoon} packages, depending on the specified algorithm.
Possible algorithms are: "xcms3", "xcms", "openms" and "envipick".
The function arguments depend on the algorithm chosen.
}
\examples{

}
\references{
\insertRef{Helmus2021}{ntsIUTA}
\insertRef{xcms1}{ntsIUTA}
\insertRef{xcms2}{ntsIUTA}
\insertRef{xcms3}{ntsIUTA}
}
