% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_003_peakPicking.R
\name{peakPicking}
\alias{peakPicking}
\title{peakPicking}
\usage{
peakPicking(
  rawData = rawData,
  param = NULL,
  removeQC = TRUE,
  refinePeaks = FALSE,
  expandRt = 1,
  minProp = 0.9,
  expandMz = 0,
  ppm = 0,
  maxPeakwidth = NULL,
  save = TRUE,
  projPath = setup$projPath,
  maxMultiProcess = TRUE
)
}
\arguments{
\item{rawData}{The \linkS4class{OnDiskMSnExp} object generated by the \code{\link{importRawData}} function.}

\item{param}{The parameters for the choosen peak picking method.
See documentation of \code{\link[xcms]{chromatographic-peak-detection}} for more information.}

\item{removeQC}{Logical, set to \code{TRUE} to remove replicate groups in the \code{rawData} named 'QC'.}

\item{refinePeaks}{Logical, set to \code{TRUE} to refine peaks.}

\item{expandRt}{Time (in seconds) to expand the peak width to merge neighboring peaks.}

\item{minProp}{The proportion (between 0 and 1) representing the proporion of intensity to be required for peaks to be joined.}

\item{expandMz}{Decreases the minimum and increases the maximum \emph{m/z} by a fixed given value.}

\item{ppm}{Decreases the minimum and increases the maximum \emph{m/z} by a fixed given value, in ppm.}

\item{maxPeakwidth}{The maximum width allowed for a peak, in seconds.}

\item{save}{Logical, set to \code{TRUE} to save the generated list of \code{XCMSnExp} objects in the disk.}

\item{projPath}{The \code{projPath} directory as defined in the \code{setup} object.}

\item{maxMultiProcess}{Logical, set to \code{TRUE} to enable max parallel processing. Changes the number of workers to the maximum available.}
}
\value{
A \code{list} of \linkS4class{XCMSnExp} objects with length equal to the number of replicate groups defined in the \code{setup}.
The \code{list} is named with the name given to the sample replicate groups.
}
\description{
Extracts chromatographic peaks from centroided \linkS4class{OnDiskMSnExp} and returns a \code{list} of \linkS4class{XCMSnExp}
objects for each replicate sample group as defined in the \code{\link{setupProject}} function.
The \linkS4class{XCMSnExp} objects can be conconated via \code{c()} function of the \pkg{xcms} package.
We separate the peaks from each replicate group to facilitate workflows that include multi-project cross-analysis.
The peak picking uses the function \code{\link[xcms]{chromatographic-peak-detection}} from the \pkg{xcms} package.
Different methods for peak picking can be used. For more information, see documentation of \code{\link[xcms]{chromatographic-peak-detection}}.
}
\note{
See parameters in \code{?refineChromPeaks} from \code{xcms} for more information when \code{refinePeaks} is set to \code{TRUE}.
}
\examples{



}
\references{
\insertRef{MSnbase1}{ntsIUTA}
\insertRef{MSnbase2}{ntsIUTA}
\insertRef{xcms1}{ntsIUTA}
\insertRef{xcms2}{ntsIUTA}
\insertRef{xcms3}{ntsIUTA}
}
