% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_003_peakPicking.R
\name{peakPicking}
\alias{peakPicking}
\title{peakPicking}
\usage{
peakPicking(obj = NULL, algorithm = NULL, param = NULL, save = TRUE)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object.}

\item{algorithm}{The algorithm to use for peak picking.}

\item{param}{List of arguments for the specified algorithm.
See \code{\link[patRoon]{findFeatures}} for more information.}

\item{save}{Logical, set to \code{TRUE} to save updated
\linkS4class{ntsData} object in the \strong{rdata} folder.
Note that \code{TRUE} overwrites the existing \linkS4class{ntsData} object.
Optionally, a character string can be given instead of \code{TRUE}
to be used as file name, avoiding overwriting.}
}
\value{
An \linkS4class{ntsData} object containing a \linkS4class{features}
object in the slot \code{patdata}.
}
\description{
Finds chromatographic peaks from centroided MS in the
samples of a given \linkS4class{ntsData} object.
The peak picking uses the \pkg{patRoon} package and
the following algorithms are possible:
"xcms3", "xcms", "openms", "envipick", "sirius", "kpic2", "safd".
The parameters depend on the algorithm chosen.
See ?\pkg{patRoon} for further information.
}
\references{
\insertRef{Helmus2021}{ntsIUTA}
\insertRef{xcms1}{ntsIUTA}
\insertRef{xcms2}{ntsIUTA}
\insertRef{xcms3}{ntsIUTA}
}
