% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotAlignment.R
\name{plotAlignment}
\alias{plotAlignment}
\title{plotAlignment}
\usage{
plotAlignment(x = features)
}
\arguments{
\item{x}{The \linkS4class{XCMSnExp} object obtained from the \code{\link{peakGrouping}} function.}
}
\value{
A plot with the retention time alignment differences for each samples.
}
\description{
Plots the results from the retention time alignment across samples.
}
