% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_009_qualityControl.R
\name{plotCheckQC}
\alias{plotCheckQC}
\title{plotCheckQC}
\usage{
plotCheckQC(obj, rtWindow = 30, ppm = 15)
}
\arguments{
\item{obj}{obj An \linkS4class{ntsData} object, containing QC samples.}

\item{rtWindow}{The time window to show the time deviation.}

\item{ppm}{The ppm window to show the mass deviation.}
}
\value{
A plot of the QC check results.
}
\description{
plotCheckQC
}
