% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotFeatures.R
\name{plotFeatures}
\alias{plotFeatures}
\title{plotFeatures}
\usage{
plotFeatures(
  x = patData,
  fileIndex = NULL,
  features = NULL,
  mz = NULL,
  ppm = 5,
  rt = NULL,
  rtWindow = 1,
  rtUnit = "min",
  title = NULL,
  plotBy = c("samples")
)
}
\arguments{
\item{x}{An \linkS4class{XCMSnExp} or \linkS4class{featureGroups} object with one or more files and grouped peaks.}

\item{fileIndex}{The index of the file/s to extract the centroids or profile data.}

\item{features}{The identifier of the features of interest.}

\item{mz}{Optional target \emph{m/z} to find features using the mass deviation specified by the \code{ppm} parameter.}

\item{ppm}{The mass deviation to extract the features when \code{mz} is specified.}

\item{rt}{The retention time in minutes or seconds, depending on the defined \code{rtUnit}, see below. Only used when \code{mz} is specified.}

\item{rtWindow}{The time deviation to collect features. The time unit is the defined by \code{rtUnit}.
A time interval can be given with a length 2 vector, defining the minimum and maximum retention time.}

\item{rtUnit}{Possible entries are \code{min} or \code{sec}. The default is \code{min}.}

\item{title}{The title for the plot, optional.}

\item{plotBy}{The grouping for the plot. Possible are "samples", "replicates" and "features"
for ploting by individual samples, relpicate samples or features, respectively. The default is "samples".}
}
\value{
An iterative plot.
}
\description{
Iterative plot for features from an \linkS4class{XCMSnExp} or \linkS4class{featureGroups} containing grouped peaks.
}
\examples{



}
