% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{plotFeatures}
\alias{plotFeatures}
\title{plotFeatures}
\usage{
plotFeatures(object, ...)
}
\arguments{
\item{object}{An object with features for plotting the respective peaks.}

\item{...}{Other method dependent parameters.}
}
\value{
A plot with chromatographic peaks, including integrated areas,
belonging to given features.
}
\description{
plotFeatures
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
