% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{plotPeaks}
\alias{plotPeaks}
\title{plotPeaks}
\usage{
plotPeaks(object, ...)
}
\arguments{
\item{object}{An object to with EICs for plotting chromatographic peaks.}

\item{...}{Other method dependent parameters.}
}
\value{
A plot with chromatographic peaks, including integrated areas.
}
\description{
plotPeaks
}
\seealso{
See method details in the \linkS4class{ntsData} help page.
}
