% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plotRawChrom.R
\name{plotRawChrom}
\alias{plotRawChrom}
\title{plotRawChrom}
\usage{
plotRawChrom(
  x = rawData,
  fileIndex = NULL,
  mz = NULL,
  ppm = 20,
  rt = NULL,
  rtWindow = NULL,
  rtUnit = "min",
  msLevel = 1,
  type = "tic"
)
}
\arguments{
\item{x}{An \linkS4class{OnDiskMSnExp} object with one or more files.}

\item{fileIndex}{The index of the file/s to extract the centroids or profile data.}

\item{mz}{Optional target \emph{m/z} to obtain an extracted ion chromatogram (EIC).}

\item{ppm}{The mass deviation to extract the data for the EIC in \code{ppm}.}

\item{rt}{The retention time in minutes or seconds, depending on the defined \code{rtUnit}, see below.}

\item{rtWindow}{The time deviation to collect centroids or profile data. The time unit is the defined by \code{rtUnit}.
A time interval can be given with a length 2 vector, defining the minimum and maximum retention time.}

\item{rtUnit}{Possible entries are \code{min} or \code{sec}. The default is \code{sec}.}

\item{msLevel}{The MS level to extract the data. POssible values are 1 or 2.}

\item{type}{The type of chromatogram. Possible entries are "bpc" of base peak chromatogram or "tic" for total ion chromatogram.
The default is "tic".}
}
\value{
An iterative
}
\description{
Plot BPC or TIC chromatograms of an \linkS4class{OnDiskMSnExp} object using \pkg{plotly} to procude an iteractive plot.
}
\examples{



}
