% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_001_makeSetup.R
\name{removeDuplicateNames}
\alias{removeDuplicateNames}
\title{removeDuplicateNames}
\usage{
removeDuplicateNames(newJoint)
}
\arguments{
\item{newJoint}{Original and new samples \code{data.frame} objects collated by \code{rbind}.}
}
\value{
Returns the same \code{data.frame} but with duplicate sample names
edited as duplicate sample names are not allowed.
}
\description{
Removes duplicate sample names to avoid problems during further analysis.
}
