% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_001_makeSetup.R
\name{saveObject}
\alias{saveObject}
\title{saveObject}
\usage{
saveObject(projPath = getwd(), polarity = "positive", ...)
}
\arguments{
\item{projPath}{The project directory. The default is \code{base::getwd()}.}

\item{polarity}{The polarity of the project. POssible values are \emph{positive} or \emph{negative}.}

\item{...}{The object to be save can be passed as it is.
For instance, for the sampleInfo \code{data.frame} it should be \code{sampleInfo = sampleInfo}
so that the object name is used as file name. For the function \code{setupProject} the date can be given
as \code{date = date} after the sampleInfo to be stored and added to the rds object names.}
}
\value{
Saves the given object in the rData folder of the project.
}
\description{
Saves project objects in the rData folder,
which is created when the \code{save} argument is set to \code{TRUE}.
}
