% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_001_makeSetup.R
\name{saveObject}
\alias{saveObject}
\title{saveObject}
\usage{
saveObject(format = "rds", filename = NULL, path = getwd(), ...)
}
\arguments{
\item{format}{The saving format. The default is \code{rds},
which is saved in rdata folder.
Other formats are saved in the results folder.}

\item{filename}{A character string with the desired RDS file name.}

\item{path}{The project path to save the \code{obj}.
The default is the working directory as obtained by \code{getwd()}.
Not needed if \code{obj} is of class \linkS4class{ntsData}
since the path is taken from the slot \code{path}.}

\item{...}{An object to save.}
}
\description{
Saves project objects as RDS in the rdata folder (the default)
or as other formats in the results folder, when specified.
If not present, the folders are created.
}
