% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_010_suspectScreening.R
\name{screenSuspectsFromCSV}
\alias{screenSuspectsFromCSV}
\title{screenSuspectsFromCSV}
\usage{
screenSuspectsFromCSV(
  obj,
  name = NULL,
  suspects = utils::read.csv(base::file.choose()),
  ppm = 5,
  rtWindow = 30,
  adduct = NULL,
  excludeBlanks = TRUE,
  withMS2 = TRUE,
  MS2param = MS2param()
)
}
\arguments{
\item{obj}{An \linkS4class{ntsData} object with features
to preform suspect screening.}

\item{name}{An optional string (without spaces) to be used as identifier for the
suspect screening entry in the \code{workflows} slot of the \linkS4class{ntsData} object.}

\item{suspects}{A \code{data.frame} or a location for a csv
with details for each suspect compound.
See details for more information about the required data.frame structure.}

\item{ppm}{The mass deviation, in ppm, to screen for the suspects.
The default is 5 ppm.}

\item{rtWindow}{The retention time deviation, in seconds,
to screen for the QC target standards. The default is 30 seconds.}

\item{adduct}{The adduct class for screening suspects. The default is \code{NULL},
which uses the polarity of the \code{obj} or the \code{adduct} class defined
in the suspects data frame. See details for more information.}

\item{excludeBlanks}{Logical, set to \code{TRUE} to ignore replicate groups
assigned as blanks in the \code{samples} slot of the \code{obj}.}

\item{withMS2}{Logical, set to \code{TRUE} for using confirmation via MS2.}

\item{MS2param}{An \linkS4class{MS2param} object with parameters for MS2 extraction.}
}
\value{
A data.frame with the suspect screening results.
}
\description{
Method to perform suspect screening from a given
list of candidates in a csv file. The method uses the S4 method
\code{screenSuspects} from \pkg{patRoon}.
}
\details{
The \code{suspects} data.frame should follow the template requires as
obtained via \code{getScreeningListTemplate()}.
Add other details of the template.
}
\examples{

}
\references{
\insertRef{Helmus2021}{ntsIUTA}
}
