% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_001_setupProject.R
\name{setupProject}
\alias{setupProject}
\title{setupProject}
\usage{
setupProject(
  path = utils::choose.dir(getwd(), "Select or create a project folder"),
  title = NA_character_,
  description = NA_character_,
  date = Sys.Date(),
  convertFiles = FALSE,
  convertFrom = NULL,
  convertToCentroid = TRUE,
  replicates = NULL,
  polarity = "positive",
  method = NA_character_,
  save = TRUE,
  makeNewProject = FALSE
)
}
\arguments{
\item{path}{The directory for project setup. The default is the \code{utils::choose.dir()} to select or create a folder.
Note that the function will look into the folder or subfolders for mzML and/or mzXML files
or specified raw files when \code{convertFiles} is set to \code{TRUE}.}

\item{title}{A character string with the project title.}

\item{description}{A character string with the project description.}

\item{date}{The project date.}

\item{convertFiles}{Logical, set to \code{TRUE} for converting vendor MS files to mzML.
The default is \code{FALSE} for no conversion even if vendor MS files are present.}

\item{convertFrom}{The name of the vendor format to be found and converted.
Possible values are: \emph{thermo}, \emph{bruker}, \emph{agilent}, \emph{ab} (from AB Sciex) and \emph{waters}.}

\item{convertToCentroid}{Logical, set to \code{TRUE} to centroid profile data when converting to mzML. The default is \code{TRUE}.}

\item{replicates}{Optional character vector with the same length as the number of MS files in the project folder
with the identifier/name of the sample replicate group for each file.
The default is \code{NULL} which attempts  to group the samples using
the name of the MS files. The groups can always be edited using
the class method \link{sampleGroups<-}.
See \code{?"sampleGroups<-"} for more information.}

\item{polarity}{A vector specifying the ionization polarity of the files.
Possible values are \emph{positive}, \emph{negative} or \emph{both} for positive, negative or both modes, respectively.
The default is \code{positive} and will be filled automatically for all found MS files.
When files from different polarities are present, the polarity of each file must be specified using the method X
or changing it manually in the \code{samples} slot of the \linkS4class{ntsData}.}

\item{method}{A character vector with the method used for acquiring the MS data.
When different methods were used, the method should be given per file.
Note that the character vector must have the same length as the number of mzML/mzXML files in the project folder.}

\item{save}{Logical, set to \code{TRUE} to save the \linkS4class{ntsData} object in the \strong{rData} folder.
If \strong{rData} folder does not exist in given \code{path}, the folder will be created.}

\item{makeNewProject}{Logical, set to TRUE to create an R project in the given \code{path} and open a new R session.}
}
\value{
An S4 class object named \linkS4class{ntsData} containing initialisation data for the NTS project.
}
\description{
\link{setupProject} takes a given path to create a project for Non-target screening (NTS).
mzML and/or mzXML files in the path are directly added to the project.
Yet, the function \code{\link{addFiles}} can be used to add files or extra files to the project.
Setting \code{createRproject} to \code{TRUE} will create and open an R project in the selected or given folder.
When \code{convertFiles} is \code{TRUE} the function uses the function \code{\link{mzMLconverter}} to automatically convert
specified MS files in the path to mzML and add them to the project.
}
\details{
The \linkS4class{ntsData} object can be edited using the class methods.
See \code{?"ntsData-class"} for more information about the slots and methods.
}
\examples{
path <-  system.file(package = "ntsIUTA", dir = "extdata")
object <- setupProject(path = path, save = FALSE)
object

}
