% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{suspectList-class}
\alias{suspectList-class}
\title{suspectList}
\value{
An \linkS4class{suspectList} object to be used within screening
workflows of \pkg{ntsIUTA}.
}
\description{
suspectList
}
\section{Slots}{

\describe{
\item{\code{path}}{The path of the csv file in disk.}

\item{\code{data}}{The data.frame listing all the suspects of interest.}

\item{\code{length}}{The number of compounds in the suspect list.}

\item{\code{comment}}{Optional comment for the suspect list.}
}}

